Spatial autocorrelation estimator

The approximate-profile likelihood estimator for the spatial autocorrelation parameter from a simultaneous autoregressive (SAR) model. Note, the APLE approximation is quite unreliable when the number of observations is large.

aple(x, w, digits = 3)

Arguments

x
Numeric vector of values, length n. This will be standardized internally with scale(x).

w
An n x n row-standardized spatial connectivity matrix. See shape2mat.

digits
Number of digits to round results to.

Source

Li, Honfei and Calder, Catherine A. and Cressie, Noel (2007). Beyond Moran's I: testing for spatial dependence based on the spatial autoregressive model. Geographical Analysis: 39(4): 357-375.

Value

the APLE estimate.

Details

To check reliability, the APLE can be compared to an estimate of the spatial autocorrelation parameter from an intercept-only SAR model.

See also

Examples

library(sf)

#> Linking to GEOS 3.8.0, GDAL 3.0.4, PROJ 6.3.1

data(georgia)
w <- shape2mat(georgia, "W")
x <- georgia$ICE
aple(x, w)

#> Row standardizing connectivity matrix

#> [1] 0.718