Auto-Gaussian family for CAR models

create a family object for the auto-Gaussian CAR or SAR specification

auto_gaussian(type)

Arguments

type

Optional; either "CAR" for conditionally specified auto-model or "SAR" for the simultaneously specified auto-model. The type is added internally by stan_car or stan_sar when needed.

Value

An object of class family

See also

Examples

cp = prep_car_data(shape2mat(georgia))
#> Range of permissible rho values:  -1.661134 1 
fit <- stan_car(log(rate.male) ~ 1,
                data = georgia,
                car_parts = cp,
                family = auto_gaussian(),
                chains = 2, iter = 700) # for speed only
#> 
#> *Setting prior parameters for intercept
#> Distribution: normal
#>   location scale
#> 1     -4.2     5
#> 
#> *Setting prior for CAR scale parameter (car_scale)
#> Distribution: student_t
#>   df location scale
#> 1 10        0     3
#> 
#> *Setting prior for CAR spatial autocorrelation parameter (sar_rho)
#> Distribution: uniform
#>   lower upper
#> 1  -1.7     1
#> 
#> SAMPLING FOR MODEL 'foundation' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000125 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.25 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 1: Iteration: 351 / 700 [ 50%]  (Sampling)
#> Chain 1: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.288 seconds (Warm-up)
#> Chain 1:                0.149 seconds (Sampling)
#> Chain 1:                0.437 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'foundation' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 7.3e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.73 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:   1 / 700 [  0%]  (Warmup)
#> Chain 2: Iteration: 351 / 700 [ 50%]  (Sampling)
#> Chain 2: Iteration: 700 / 700 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.223 seconds (Warm-up)
#> Chain 2:                0.166 seconds (Sampling)
#> Chain 2:                0.389 seconds (Total)
#> Chain 2: 
print(fit)
#> Spatial Model Results 
#> Formula: log(rate.male) ~ 1
#> <environment: 0x55d5d3939d58>
#> Spatial method (outcome):  CAR 
#> Likelihood function:  auto_gaussian 
#> Link function:  identity 
#> Residual Moran Coefficient:  -0.1924671 
#> WAIC:  4.67 
#> Observations:  159 
#> Data models (ME): none
#> Inference for Stan model: foundation.
#> 2 chains, each with iter=700; warmup=350; thin=1; 
#> post-warmup draws per chain=350, total post-warmup draws=700.
#> 
#>             mean se_mean    sd   2.5%    25%    50%    75%  97.5% n_eff  Rhat
#> intercept -4.157   0.003 0.058 -4.271 -4.194 -4.156 -4.123 -4.039   361 0.998
#> car_rho    0.855   0.004 0.088  0.654  0.810  0.868  0.922  0.981   530 0.999
#> car_scale  0.526   0.002 0.034  0.464  0.501  0.527  0.548  0.597   386 0.998
#> 
#> Samples were drawn using NUTS(diag_e) at Wed Apr 26 19:40:27 2023.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).