Edge list

Creates a list of connected nodes following the graph representation of a spatial connectivity matrix.

edges(C, unique_pairs_only = TRUE)

Arguments

C
A connectivity matrix where connection between two nodes is indicated by non-zero entries.

unique_pairs_only
By default, only unique pairs of nodes (i, j) will be included in the output.

Value

Returns a data.frame with three columns. The first two columns (node1 and node2) contain the indices of connected pairs of nodes; only unique pairs of nodes are included (unless `unique_pairs_only = FALSE`). The third column (weight) contains the corresponding matrix element, C[node1, node2].

Details

This is used internally for stan_icar and it is also helpful for creating the scaling factor for BYM2 models fit with stan_icar.

See also

Examples

data(sentencing)
C <- shape2mat(sentencing)
nbs <- edges(C)
head(nbs)
#>   node1 node2 weight
#> 1     1     3      1
#> 2     1     7      1
#> 3     1     8      1
#> 4     1    20      1
#> 5     1    24      1
#> 6     1    28      1

## similar to:
head(Matrix::summary(C))
#> 47 x 47 sparse Matrix of class "ngCMatrix", with 208 entries 
#>    i j
#> 1  3 1
#> 2  7 1
#> 3  8 1
#> 4 20 1
#> 5 24 1
#> 6 28 1
head(Matrix::summary(shape2mat(georgia, "W")))
#> 159 x 159 sparse Matrix of class "dgCMatrix", with 860 entries 
#>     i j         x
#> 1  23 1 0.1666667
#> 2  58 1 0.1428571
#> 3  59 1 0.1666667
#> 4 131 1 0.2000000
#> 5 148 1 0.1428571
#> 6 159 1 0.1250000