Expected dimensions of an eigenvector spatial filter

Provides an informed guess for the number of eigenvectors required to remove spatial autocorrelation from a regression. This is used internally for stan_esf; the result can be used to set the prior scale parameter for the global shrinkage parameter in the regularized horseshoe prior. A smaller value of `p0` leads to a more sparse specification.

exp_pars(formula, data, C)

Arguments

formula
Model formula.

data
The data used to fit the model; must be coercible to a dataframe for use in model.matrix.

C
An N x N binary connectivity matrix.

Source

Chun, Y., D. A. Griffith, M. Lee and P. Sinha (2016). Eigenvector selection with stepwise regression techniques to construct eigenvector spatial filters. *Journal of Geographical Systems*, 18(1), 67-85. doi: 10.1007/s10109-015-0225-3 .

Donegan, C., Y. Chun and A. E. Hughes (2020). Bayesian estimation of spatial filters with Moran’s Eigenvectors and hierarchical shrinkage priors. *Spatial Statistics*. doi: 10.1016/j.spasta.2020.100450 .

Piironen, J and A. Vehtari (2017). Sparsity information and regularization in the horseshoe and other shrinkage priors. In *Electronic Journal of Statistics*, 11(2):5018-5051. doi: 10.1214/17-EJS1337SI .

Value

Returns a numeric value representing the expected number of eigenvectors required to estimate a spatial filter (i.e. number of non-zero or 'large' coefficients).

Details

Following Chun et al. (2016), the expected number of eigenvectors required to remove residual spatial autocorrelation from a model is an increasing function of the degree of spatial autocorrelation in the outcome variable and the number of links in the connectivity matrix.

See also

Examples

C <- shape2mat(georgia, "B")
c(p0 = exp_pars(log(rate.male) ~ college, georgia, C))
#> p0 
#>  2 
if (FALSE) {
 fit <- stan_esf(log(rate.male) ~ college, data = georgia, p0 = p0, iter = 1e3)
}