Georgia all-cause, sex-specific mortality, ages 55-64, years 2014-2018

A simple features (sf) object for Georgia counties with sex- and age-specific deaths and populations at risk (2014-2018), plus select estimates (with standard errors) of county characteristics. Standard errors of the ICE were calculated using the Census Bureau's variance replicate tables.

georgia

Format

A simple features object with county geometries and the following columns:

GEOID

Six digit combined state and county FIPS code

NAME

County name

ALAND

Land area

AWATER

Water area

population

Census Bureau 2018 county population estimate

white

Percent White, ACS 2018 five-year estimate

black

Percent Black, ACS 2018 five-year estimate

hisp

Percent Hispanic/Latino, ACS 2018 five-year estimate

ai

Percent American Indian, ACS 2018 five-year estimate

deaths.male

Male deaths, 55-64 yo, 2014-2018

pop.at.risk.male

Population estimate, males, 55-64 yo, years 2014-2018 (total), ACS 2018 five-year estimate

pop.at.risk.male.se

Standard error of the pop.at.risk.male estimate

deaths.female

Female deaths, 55-64 yo, 2014-2018

pop.at.risk.female

Population estimate, females, 55-64 yo, years 2014-2018 (total), ACS 2018 five-year estimate

pop.at.risk.female.se

Standard error of the pop.at.risk.female estimate

ICE

Index of Concentration at the Extremes

ICE.se

Standard error of the ICE estimate, calculated using variance replicate tables

income

Median household income, ACS 2018 five-year estimate

income.se

Standard error of the income estimate

college

Percent of the population age 25 or higher than has a bachelors degree of higher, ACS 2018 five-year estimate

college.se

Standard error of the college estimate

insurance

Percent of the population with health insurance coverage, ACS 2018 five-year estimate

insurance.se

Standard error of the insurance estimate

rate.male

Raw (crude) age-specific male mortality rate, 2014-2018

rate.female

Raw (crude) age-specific female mortality rate, 2014-2018

geometry

simple features geometry for county boundaries

Source

Centers for Disease Control and Prevention, National Center for Health Statistics. Underlying Cause of Death 1999-2018 on CDC Wonder Online Database. 2020. Available online: http://wonder.cdc.gov (accessed on 19 October 2020).

Donegan, Connor and Chun, Yongwan and Griffith, Daniel A. (2021). ``Modeling community health with areal data: Bayesian inference with survey standard errors and spatial structure.'' Int. J. Env. Res. and Public Health 18 (13): 6856. DOI: 10.3390/ijerph18136856 Data and code: https://github.com/ConnorDonegan/survey-HBM.

Kyle Walker and Matt Herman (2020). tidycensus: Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames. R package version 0.11. https://CRAN.R-project.org/package=tidycensus

US Census Bureau. Variance Replicate Tables, 2018. Available online: https://www.census.gov/programs-surveys/acs/data/variance-tables.2018.html (accessed on 19 October 2020).

Examples

data(georgia)
head(georgia)

library(sf)
plot(georgia[,'rate.female'])