geostan_fit methods

Methods for fitted geostan models: extract residuals, fitted values, posterior predictive distribution or spatial component from a spatial regression model; extract samples from the posterior distribution; print regression results; plot posterior distributions.

# S3 method for geostan_fit
print(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), digits = 3, ...)

# S3 method for geostan_fit
plot(x, pars, plotfun = "dens", ...)

# S3 method for geostan_fit
as.matrix(x, ...)

# S3 method for geostan_fit
as.data.frame(x, ...)

# S3 method for geostan_fit
as.array(x, ...)

# S3 method for geostan_fit
residuals(object, summary = TRUE, ...)

# S3 method for geostan_fit
fitted(object, summary = TRUE, ...)

# S3 method for geostan_fit
spatial(object, summary = TRUE, ...)

Arguments

x
A fitted model object of class geostan_fit.

probs
Argument passed to quantile; which quantiles to calculate and print.

digits
number of digits to print

...
additional arguments.

pars
parameters to include; a character string or vector of parameter names.

plotfun
Argument passed to rstan::plot. Options include "hist", "trace", and "dens".

object
A fitted model object of class geostan_fit.

summary
Logical; should the values be summarized with the mean, standard deviation and quantiles (probs = c(.025, .2, .5, .8, .975)) for each observation? Otherwise a matrix containing samples from the posterior distribution at each observation is returned.

Value

Methods residuals, fitted, spatial return a matrix containing all samples for each observation if summary = FALSE, else if summary = TRUE a data.frame containing a summary of the posterior distribution at each observation (of, respectively, residuals, fitted values, or the spatial trend).

plot returns a ggplot object that can be customized using the ggplot2 package.

as.matrix, as.data.frame, as.array return samples from the (joint) posterior distribution of parameters in the format corresponding to their names. The pars argument is used to return samples from only a subset of parameters.

See also

Examples

if (FALSE) {
library(ggplot2)
library(sf)
data(ohio)
fit <- stan_esf(gop_growth ~ historic_gop + log(pop_density),
               data = ohio,
               scalex = TRUE,
               C = shape2mat(ohio),
               refresh = 0 # less printing
)

# print and plot results
print(fit)
plot(fit)

beta.samples <- as.matrix(fit, pars = "beta")
beta.mean <- apply(beta.samples, 2, mean)
print(beta.mean)

# extract residuals and fitted values
res <- residuals(fit)
res <- resid(fit) 
head(res)
f <- fitted(fit)
plot(f$mean, res$mean, xlab = "fitted values", ylab = "residuals")
mat <- resid(fit, summary = FALSE)
dim(mat)

# extract and plot the posterior mean of the spatial filter
sf <- spatial(fit, summary = TRUE)
head(sf)
ohio$sf <- sf$mean
ggplot(ohio) +
  geom_sf(aes(fill = sf)) +
  scale_fill_gradient2()  
}