Local Moran's I

A local indicator of spatial association (lisa).

lisa(x, w, type = FALSE)

Arguments

x
Numeric vector

w
An n x n spatial connectivity matrix. See shape2mat. This will automatically be row-standardized!

type
Return the type of association also (High-High, Low-Low, High-Low, and Low-High)? Defaults to FALSE.

Source

Anselin, Luc. "Local indicators of spatial association—LISA." Geographical analysis 27, no. 2 (1995): 93-115.

Value

If type = FALSE a numeric vector of lisa values for exploratory analysis of local spatial autocorrelation. If type = TRUE, a data.frame with columns zi (the lisa value) and type.

Details

The values will be standardized with scale(x) first and w will be row-standardized. Then the LISA is the product of a value with its mean surrounding value. These are for exploratory analysis and model diagnostics. The function uses Equation 7 from Anselin (1995).

An above-average value (i.e. positive z-value) with positive mean spatial lag is of type "High-High"; a low value surrounded by high values is of type "Low-High", and so on.

See also