Ohio Presidential election results and county characteristics

A simple features spatial dataset containing county attributes and Presidential election data for Ohio. This is a processed version of the MIT "County Presidential Election Returns 2000-2016" data set augmented with data from the American Community Survey and Bureau of Labor Statistics.

ohio

Format

A simple feature collection including the following attributes:

GEOID

Six digit combined state and county FIPS code

county

County name

gop_growth

Change in the Republican vote share from historic (2000 - 2012) average vote share to 2016 vote share (i.e. (trump_2016/total_2016) - historic_gop). Data only includes Democratic and Republican votes (i.e. two-party vote share).

historic_gop

Average Republican share of all major party votes, 2000 to 2012

trump_2016

Number of votes for Donald Trump in 2016

total_2016

Total number of Democratic and Republican votes in 2016

population

ACS 2016 5 year population estimate

pop_density

Population per square mile (population / ALAND)

college_educated

ACS 2016 5 year estimate of the percent of the population 25 and older with a bachelors degree or higher

college_educated.se

Standard error for college_educated

white_nonhispanic

ACS 2016 5 year estimate of the percent of non-hispanic whites in the population

white_nonhispanic.se

Standard error for white_nonhispanic

unemployment

Local area unemployment estimate from the Bureau of Labor Statistics, 2016

unemployment.acs

Civilian unemployment rate estimate from the American Community Survey, 2016

unemployment.acs.se

Standard error for unemployment.acs

geometry

County boundaries (polygons) in simple features format

Source

Donegan, C., Y. Chun and A. E. Hughes (2020). Bayesian Estimation of Spatial Filters with Moran’s Eigenvectors and Hierarchical Shrinkage Priors. Spatial Statistics. https://doi.org/10.1016/j.spasta.2020.100450

MIT Election Data and Science Lab, 2018, "County Presidential Election Returns 2000-2016", https://doi.org/10.7910/DVN/VOQCHQ, Harvard Dataverse, V1, UNF:6:ZaxsDvp6RsFitno8ZYlK7w== [fileUNF]

US Bureau of Labor Statistics, Local Force By County, 2016 Annual Averages. https://www.bls.gov/lau/#cntyaa

US Census Bureau, 2016. American Community Survey. Tables DP02, DP03, DP05 5-year estimates.

Kyle Walker, 2018. Tidycensus. https://walkerke.github.io/tidycensus/index.html

Details

The Bureau of Labor Statistics unemployment estimates are based on the ACS estimates as well as additional information. The BLS estimates were used for the analysis in Donegan et al. (2020). The ACS estimates have standard errors (the BLS estimates do not).

Examples

if (FALSE) {
library(sf)
data(ohio)
}