Prepare data for Stan CAR model

Prepare data for Stan CAR model

prep_car_data(
  A,
  style = c("ACAR", "WCAR", "DCAR"),
  lambda = FALSE,
  cmat = TRUE,
  d,
  k = 1
)

Arguments

A
binary adjacency matrix

style
specification for the connectivity matrix (C) and conditional variances (M)

lambda
If TRUE, return eigenvalues required for calculating the log determinant of the precision matrix and for determining the range of permissible values of rho. These will also be printed with a message if lambda = TRUE.

cmat
Return the full matrix C if TRUE.

d
distance matrix. Non-neighboring values (as indicated by A) will be set to zero internally.

k
inverse distances will be raised to the -k power (d^-k).

Source

Cliff A, Ord J (1981). Spatial Processes: Models and Applications. Pion.

Cressie N (2015 [1993]). Statistics for Spatial Data. Revised edition. John Wiley & Sons.

Cressie N, Perrin O, Thomas-Agnan C (2005). “Likelihood-based estimation for Gaussian MRFs.” Statistical Methodology, 2(1), 1–16.

Cressie N, Wikle CK (2011). Statistics for Spatio-Temporal Data. John Wiley & Sons.

Haining RP, Li G (2020). Modelling Spatial and Spatio-Temporal Data: A Bayesian Approach. CRC Press.

Value

A list containing all of the data elements required by the Stan CAR model.

Details

The CAR model is N(Mu, Sigma), Sigma = (I - rho C)^-1 M.

The DCAR specification is distance-based, and requires the user provide a distance matrix d in addition to adjacency matrix A. The WCAR specification uses row-standardized connectivity matrix and sets conditional variances proportional to the number of neighbors. The ACAR specification is from Cressie, Perrin and Thomas-Agnon (2005); also see Cressie and Wikle (2011, p. 188).

For inverse-distance weighting schemes, see Cliff and Ord (1981); for distance-based CAR specifications, see Cressie (2015 [1993]) and Haining and Li (2020).

stan_car requires the user to provide the full C matrix (which is returned when cmat = TRUE).