Prior distributions

Prior distributions

uniform(lower, upper, variable = NULL)

normal(location = 0, scale, variable = NULL)

student_t(df = 10, location = 0, scale, variable = NULL)

gamma(alpha, beta, variable = NULL)

hs(global_scale = 1, slab_df = 10, slab_scale, variable = "beta_ev")

Value

Return value for student_t depends on the input; if no arguments are provided (specifically, if the scale parameter is missing), this will return an object of class 'family'; if at least the scale parameter is provided, `student_t` will return an object of class `prior` containing parameter values for the Student's t distribution.

Details

The prior distribution functions are used to set the values of prior parameters.

Note that users can control the values of the parameters, but the distribution (model) itself is fixed. The intercept and regression coefficients are given Gaussian prior distributions and scale parameters are assigned Student's t prior distributions. Degrees of freedom pararmeters are assigned gamma priors, and the spatial autocorrelation parameter in the CAR model, rho, is assigned a uniform prior. The horseshoe (`hs`) model is used by stan_esf.

Note that the `variable` argument is used internally by `geostan`, and any user provided values will be ignored.

Examples

prior <- list()
prior$beta <- normal(c(0, 0), c(1, 1))
prior$intercept <- normal(-5, 3)
if (FALSE) {
fit <- stan_glm(deaths.male ~ offset(log(pop.at.risk.male)) + ICE + college,
                re = ~ GEOID,
                data = georgia,
                family = poisson(),
                prior = prior,
                prior_only = TRUE)
plot(fit)
}

ME <- list()
ME$se <- data.frame(insurance = georgia$insurance.se)
ME$prior <- list()
ME$prior$nu <- gamma(3, 0.2)
ME$prior$location <- normal(50, 50)
ME$prior$scale <- student_t(12, 10, 20)
if (FALSE) {
fit <- stan_glm(log(rate.male) ~ insurance, 
                data = georgia,
                ME = ME,
                prior_only = TRUE)
}