Simulate Spatially Autocorrelated Data

Given a spatial weights matrix and degree of autocorrelation, returns autocorrelated data.

sim_sar(n = 1, mu = rep(0, nrow(w)), w, rho, sigma = 1, ...)

Arguments

n
The number of samples required. Defaults to n=1 to return a k-length vector; if n>1, an n x k matrix is returned (i.e. each row will contain a sample of correlated values).

mu
A k-length vector of mean values. Defaults to a vector of zeros with length equal to nrow(W).

w
Row-standardized k x k spatial weights matrix.

rho
Spatial autocorrelation parameter in the range [-1, 1]. Typically a scalar value; otherwise a K-length numeric vector.

sigma
Scale parameter (standard deviation). Defaults to sigma = 1. Typically a scalar value; otherwise a K-length numeric vector.

...
further arguments passed to MASS::mvrnorm.

Value

If n = 1 a vector of the same length as mu, otherwise an n x length(mu) matrix with one sample in each row.

Details

Calls MASS::mvrnorm internally to draw from the multivariate normal distribution. The covariance matrix is specified following the simultaneous autoregressive (SAR) model.

See also

Examples

data(ohio)
w <- shape2mat(ohio, "W")
x <- sim_sar(w=w, rho=.8)
aple(x, w)

#> [1] 0.778