Intrinsic autoregressive models

Assign the intrinsic conditional auto-regressive (ICAR) prior model to parameters. Options include the BYM model, the BYM2 model, and a solo ICAR term.

stan_icar(
  formula,
  slx,
  re,
  data,
  type = c("icar", "bym", "bym2"),
  scale_factor = NULL,
  ME = NULL,
  C,
  family = poisson(),
  prior = NULL,
  prior_intercept = NULL,
  prior_tau = NULL,
  centerx = FALSE,
  scalex = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 4000,
  refresh = 1000,
  pars = NULL,
  control = list(adapt_delta = 0.9, max_treedepth = 15),
  silent = FALSE,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models can be specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally). These will be pre-multiplied by a row-standardized spatial weights matrix and then added (prepended) to the design matrix. If and when setting priors for beta manually, remember to include priors for any SLX terms as well.

re
If the model includes a varying intercept term alpha_re specify the grouping variable here using formula syntax, as in ~ ID. Then, alpha_re ~ N(0, alpha_tau), alpha_tau ~ Student_t(d.f., location, scale). Before using this, read the Details section and the type argument.

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

type
Defaults to "icar" (partial pooling of neighboring observations through parameter phi); specify "bym" to add a second parameter vector theta to perform partial pooling across all observations; specify "bym2" for the innovation introduced by Riebler et al. (2016). See Details for more information.

scale_factor
For the BYM2 model, optional. If missing, this will be set to a vector of ones.

ME
To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. Errors are assigned a Gaussian probability distribution and the modeled (true) covariate vector is assigned a Student's t model or, if ME$spatial = TRUE, an auto Gaussian (CAR) model. Elements of the list ME may include:

se

a dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounded

If any variables in se are bounded within some range (e.g. percentages ranging from zero to one hundred) provide a vector of zeros and ones indicating which columns are bounded. By default the lower bound will be 0 and the upper bound 100, for percentages.

bounds

A numeric vector of length two providing the upper and lower bounds, respectively, of any bounded variables.

spatial

Logical value indicating whether an auto Gaussian (i.e., conditional autoregressive (CAR)) model should be used for the covariates. Defaults to spatial = FALSE. The specification of the auto Gaussian model is fixed such that any non-zero values in C will be converted to ones.

C
Spatial connectivity matrix which will be used to construct an edge list for the ICAR model, and to calculate residual spatial autocorrelation as well as any user specified slx terms or spatial measurement error (ME) models. It will automatically be row-standardized before calculating slx terms. C must be a binary symmetric n x n matrix.

family
The likelihood function for the outcome variable. Current options are binomial(link = "logit") and poisson(link = "log").

prior
A data.frame or matrix with location and scale parameters for Gaussian prior distributions on the model coefficients. Provide two columns---location and scale---and a row for each variable in their order of appearance in the model formula. Default priors are weakly informative relative to the scale of the data.

prior_intercept
A vector with location and scale parameters for a Gaussian prior distribution on the intercept; e.g. prior_intercept = c(0, 10).

prior_tau
Set hyperparameters for the scale parameter of exchangeable random effects/varying intercepts. The random effects are given a normal prior with scale parameter alpha_tau. The latter is given a half-Student's t prior with default of 20 degrees of freedom, centered on zero and scaled to the data to be weakly informative. To adjust it use, e.g., prior_tau = c(df = 15, location = 0, scale = 5).

centerx
Should the covariates be centered prior to fitting the model? Defaults to FALSE.

scalex
Should the covariates be centered and scaled (divided by their standard deviation)? Defaults to FALSE.

prior_only
Draw samples from the prior distributions of parameters only.

chains
Number of MCMC chains to estimate.

iter
Number of samples per chain. .

refresh
Stan will print the progress of the sampler every refresh number of samples; set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model.

control
A named list of parameters to control the sampler's behavior. See stan for details. The defaults are the same rstan::stan except that adapt_delta is raised to .9 and max_treedepth = 15.

silent
If TRUE, suppress printed messages including prior specifications and Stan sampling progress (i.e. refresh=0). Stan's error and warning messages will still print.

...
Other arguments passed to sampling. For multi-core processing, you can use cores = parallel::detectCores(), or run options(mc.cores = parallel::detectCores()) first.

Source

Besag, J. (1974). Spatial interaction and the statistical analysis of lattice systems. Journal of the Royal Statistical Society: Series B (Methodological), 36(2), 192-225.

Besag, J., York, J., & Mollié, A. (1991). Bayesian image restoration, with two applications in spatial statistics. Annals of the institute of statistical mathematics, 43(1), 1-20.

Morris, M., Wheeler-Martin, K., Simpson, D., Mooney, S. J., Gelman, A., & DiMaggio, C. (2019). Bayesian hierarchical spatial models: Implementing the Besag York Mollié model in stan. Spatial and spatio-temporal epidemiology, 31, 100301.

Riebler, A., Sorbye, S. H., Simpson, D., & Rue, H. (2016). An intuitive Bayesian spatial model for disease mapping that accounts for scaling. Statistical Methods in Medical Research, 25(4), 1145-1165.

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame

diagnostic

Widely Applicable Information Criteria (WAIC) with crude measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and residual spatial autocorrelation (Moran coefficient of the residuals). Residuals are relative to the mean posterior fitted values.

stanfit

an object of class stanfit returned by rstan::stan

data

a data frame containing the model data

edges

The edge list representing all unique sets of neighbors and the weight attached to each pair (i.e., their corresponding element in the connectivity matrix C

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including ICAR component)

slx

The slx formula

re

A list with two name elements, formula and Data, containing the formula re and a data frame with columns id (the grouping variable) and idx (the index values assigned to each group).

priors

Prior specifications.

scale_params

A list with the center and scale parameters returned from the call to base::scale on the model matrix. If centerx = FALSE and scalex = FALSE then it is an empty list.

spatial

A data frame with the name of the spatial parameter ("phi" if type = "icar" else "convolution") and method (toupper(type)).

Details

The Stan code for the ICAR component of the model and the BYM2 option draws from Morris et al. (2019) with adjustments to enable non-binary weights and disconnected graph structures. The ICAR parameters are returned in the parameter vector named phi. The ICAR prior model is equivalent to a CAR model with the spatial autocorrelation parameter car_alpha equal to 1 (see stan_car). Thus the ICAR prior places high probability on a smooth spatially varying mean. Often, an observational-level random effect term theta is added to model deviations from the local mean. The combination phi + theta is known as the BYM model (Besag et al. 1991).

The ICAR prior is placed on the parameter vector phi_tilde (which is approximately on the standard normal scale); it is scaled by the scalar parameter spatial_scale. The models are specified as follows:

If type = "icar", the spatial structure is simply phi[i] = phi_tilde[i] * spatial_scale.

If type = "bym", the spatial structure phi is the same as "icar" but an additional parameter vector theta is added to perform partial pooling across all observations. theta[i] = theta_tilde[i] * theta_scale. The sum phi_tilde * spatial_scale + theta_tilde * theta_scale = phi + theta = convolution is often referred to as the ``convolved random effect.'' It is returned in the parameter vector named convolution.

For type = "bym2", phi_tilde and theta_tilde share a single scale parameter (spatial_scale) but they are combined using a mixing parameter rho. The convolution is then convolution = [sqrt(rho / scale_factor) * phi_tilde + sqrt((1 - rho)) * theta_tilde] * spatial_scale. For convenience, the model will return the convolution term and also factor it into parameters phi and theta. The actual calculation of the convolution term is specified to ensure that observations with zero neighbors are handled properly.

See also

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(sf)
library(rstan)
options(mc.cores = parallel::detectCores())
data(sentencing)

C <- shape2mat(sentencing)
log_e <- log(sentencing$expected_sents)
fit.bym <- stan_icar(sents ~ offset(log_e),
                     family = poisson(),
                     data = sentencing,
                     type = "bym",
                     C = C,
                     refresh = 0
 )

# diagnostics plot: Rhat values should all by very near 1
library(rstan)
rstan::stan_rhat(fit.bym$stanfit)
 # see effective sample size for all parameters and generated quantities
 # (including residuals, predicted values, etc.)
rstan::stan_ess(fit.bym$stanfit)
# or for a particular parameter
rstan::stan_ess(fit.bym$stanfit, "phi")
library(bayesplot)
yrep <- posterior_predict(fit.bym, samples = 95)
y <- sentencing$sents
ppc_dens_overlay(y, yrep)
}