geostan_fit methods

Methods for fitted geostan models: extract residuals, fitted values, posterior predictive distribution or spatial component from a spatial regression model; extract samples from the posterior distribution; print regression results; plot posterior distributions.

# S3 method for geostan_fit
print(
  x,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  digits = 3,
  pars = NULL,
  ...
)

# S3 method for geostan_fit
plot(x, pars, plotfun = "hist", fill = "steelblue4", ...)

# S3 method for geostan_fit
as.matrix(x, ...)

# S3 method for geostan_fit
as.data.frame(x, ...)

# S3 method for geostan_fit
as.array(x, ...)

# S3 method for geostan_fit
residuals(object, summary = TRUE, rates = TRUE, detrend = TRUE, ...)

# S3 method for geostan_fit
fitted(object, summary = TRUE, rates = TRUE, ...)

spatial(object, summary = TRUE, ...)

# S3 method for geostan_fit
spatial(object, summary = TRUE, ...)

# S3 method for geostan_fit
predict(
  object,
  newdata,
  alpha = mean(as.matrix(object, pars = "intercept")),
  center = object$x_center,
  summary = TRUE,
  type = c("link", "response"),
  ...
)

Arguments

x

A fitted model object of class geostan_fit.

probs

Argument passed to quantile; which quantiles to calculate and print.

digits

number of digits to print

pars

parameters to include; a character string (or vector) of parameter names.

...

additional arguments.

plotfun

Argument passed to rstan::plot. Options include histograms ("hist"), MCMC traceplots ("trace"), and density plots ("dens"). Diagnostic plots are also available such as Rhat statistics ("rhat"), effective sample size ("ess"), and MCMC autocorrelation ("ac").

fill

fill color for histograms and density plots.

object

A fitted model object of class geostan_fit.

summary

Logical; should the values be summarized with the mean, standard deviation and quantiles (probs = c(.025, .2, .5, .8, .975)) for each observation? Otherwise a matrix containing samples from the posterior distribution at each observation is returned.

rates

For Poisson and Binomial models, should the fitted values be returned as rates, as opposed to raw counts? Defaults to TRUE.

detrend

For CAR models with Gaussian likelihood only (auto-gaussian); if detrend = TRUE, the implicit spatial trend will be removed from the residuals. The implicit spatial trend is Trend = rho * C %*% (Y - Mu) (see stan_car). I.e., resid = Y - (Mu + Trend).

newdata

A data frame in which to look for variables with which to predict, presumably for the purpose of viewing marginal effects. Note that if the model formula includes an offset term, newdata must contain a column with the appropriate name for the offset, even though the values will be ignored (you may set all values to 1); you must use the alpha argument to include any additional terms. Note also that any spatially-lagged covariate terms will be ignored if they were provided using the slx argument. If covariates in the model were centered using the centerx argument, the predict.geostan_fit method will automatically center the predictors in newdata internally using the values stored in fit$x_center. If newdata is missing, user arguments will be passed to the fitted.geostan_fit method to return the fitted values of the model.

alpha

A single numeric value or a numeric vector with length equal to nrow(newdata); alpha serves as the intercept in the linear predictor. The default is to use the posterior mean of the intercept. Even if type = "response", this needs to be provided on the scale of the linear predictor. See Details for additional information.

center

May be a vector of numeric values or a logical scalar to pass to scale. Defaults to using object$x_center. If the model was fit using centerx = TRUE, then covariates were centered and their mean values are stored in object$x_center and the predict method will use them to automatically center newdata; if the model was fit with centerx = FALSE, then object$x_center = FALSE and newdata will not be centered.

type

By default, results from predict are on the scale of the linear predictor (type = "link")). The alternative (type = "response") is on the scale of the response variable. For example, the default return values for a Poisson model are log-rates, and using type = "response" will return the rates (by exponentiating the log-rates).

Value

Methods residuals, fitted, predict, and spatial return a matrix containing all samples for each observation if summary = FALSE, else if summary = TRUE a data.frame containing a summary of the posterior distribution at each observation (of, respectively, residuals, fitted values, predicted values, or the spatial trend). The predict method will return a data frame with a summary of results together with use-provided newdata. The predict method is designed for reviewing marginal effects of covariates. Thus, results do not include spatial trends or offset terms. To obtain the fitted values of the model (as opposed to predictions from new data), use the fitted method. For the posterior predictive distribution, see posterior_predict. plot returns a ggplot object that can be customized using the ggplot2 package. as.matrix, as.data.frame, as.array return samples from the joint posterior distribution of parameters in the format corresponding to their names. The pars argument is used to return samples from only a subset of parameters.

Details

predict.geostan_fit

The purpose of the predict method is to explore marginal effects of (combinations of) covariates. The method sets the intercept equal to its posterior mean (i.e., alpha = mean(as.matrix(object, pars = "intercept"))); the only source of uncertainty in the results is the posterior distribution of the coefficients, which can be obtained using Beta = as.matrix(object, pars = "beta").

Be aware that in non-linear models (including Poisson and Binomial models) marginal effects of each covariate are sensitive to the level of other covariates in the model. If the model includes any spatially-lagged covariates (introduced using the slx argument) or a spatial autocorrelation term, these terms will essentially be fixed at zero for the purposes of calculating marginal effects. To explore the impact of these (missing) terms, you can add their values to the linear predictor using the alpha argument.

See also

Examples

# \donttest{
data(georgia)
C <- shape2mat(georgia, style = "B")
cars <- prep_car_data(C)
georgia$income <- georgia$income/1e3

fit <- stan_car(deaths.male ~ offset(log(pop.at.risk.male)) + log(income),
                slx = ~ log(income),
                centerx = TRUE,
                car_parts = cars,
                data = georgia,
                family = poisson(),
                chains = 2, iter = 600) # for speed only

# print and plot results
print(fit)
plot(fit)

# residuals
r = resid(fit)

# fitted values 
# (Poisson model defaults to rates)
f1 = fitted(fit)
f2 = fitted(fit, rates = FALSE)

# spatial diagnostics
sp_diag(fit, georgia)

# spatial trend, county `random effects' 
sp = spatial(fit)

# posterior predictive distribution
yrep <- posterior_predict(fit, S = 65)
plot(density(yrep[1,]), col = "gray30")
for (i in 2:nrow(yrep)) lines(density(yrep[i,]), col = "gray30")
lines(density(georgia$deaths.male), col = "darkred", lwd = 2)

# array of samples; MCMC diagnostics
S.array <- as.array(fit, pars = c("intercept", "car_scale", "car_rho"))
S.monitor <- rstan::monitor(S.array, print = FALSE, warmup = 0)
head(S.monitor)

newdata <- data.frame(
    income = seq(min(georgia$income), max(georgia$income), by = 1),
    pop.at.risk.male = 1
)

p <- predict(fit, newdata, type = "response")
plot(newdata$income, p$mean * 1e3,
    type = 'l',
     main = "Deaths per 1,000",
     ylab = NA,
     xlab = "Median county income ($1,000s)")
# }