geostan_fit methods

Methods for fitted geostan models: extract residuals, fitted values, posterior predictive distribution or spatial component from a spatial regression model; extract samples from the posterior distribution; print regression results; plot posterior distributions.

# S3 method for geostan_fit
print(x, probs = c(0.025, 0.25, 0.5, 0.75, 0.975), digits = 3, ...)

# S3 method for geostan_fit
plot(x, pars, plotfun = "hist", fill = "steelblue4", ...)

# S3 method for geostan_fit
as.matrix(x, ...)

# S3 method for geostan_fit
as.data.frame(x, ...)

# S3 method for geostan_fit
as.array(x, ...)

# S3 method for geostan_fit
residuals(object, summary = TRUE, rates = TRUE)

# S3 method for geostan_fit
fitted(object, summary = TRUE, rates = TRUE, ...)

# S3 method for geostan_fit
spatial(object, summary = TRUE, ...)

Arguments

x
A fitted model object of class geostan_fit.

probs
Argument passed to quantile; which quantiles to calculate and print.

digits
number of digits to print

...
additional arguments.

pars
parameters to include; a character string or vector of parameter names.

plotfun
Argument passed to rstan::plot. Options include histograms ("hist"), MCMC traceplots ("trace"), and density plots ("dens"). Diagnostic plots are also available such as Rhat statistics ("rhat"), effective sample size ("ess"), and MCMC autocorrelation ("ac").

fill
fill color for histograms and density plots.

object
A fitted model object of class geostan_fit.

summary
Logical; should the values be summarized with the mean, standard deviation and quantiles (probs = c(.025, .2, .5, .8, .975)) for each observation? Otherwise a matrix containing samples from the posterior distribution at each observation is returned.

rates
For Poisson and Binomial models, should the fitted values be returned as rates, as opposed to raw counts? Defaults to `TRUE`.

Value

Methods residuals, fitted, spatial return a matrix containing all samples for each observation if summary = FALSE, else if summary = TRUE a data.frame containing a summary of the posterior distribution at each observation (of, respectively, residuals, fitted values, or the spatial trend).

plot returns a ggplot object that can be customized using the ggplot2 package.

as.matrix, as.data.frame, as.array return samples from the joint posterior distribution of parameters in the format corresponding to their names. The pars argument is used to return samples from only a subset of parameters.

See also

Examples

if (FALSE) {
library(ggplot2)
library(sf)
data(sentencing)

# spatial weights matrix with binary coding scheme
C <- shape2mat(sentencing, style = "B")

# log-expected number of sentences
## expected counts are based on county racial composition and mean sentencing rates
log_e <- log(sentencing$expected_sents)

# fit spatial Poisson model with unstructured 'random effects'
fit <- stan_glm(sents ~ offset(log_e),
                   re = ~ name,
                   family = poisson(),
                   data = sentencing,
                   C = C,
                   refresh = 0
)

# print and plot results
print(fit)
plot(fit)

# residuals
r = resid(fit)

# fitted values
f = fitted(fit)

# spatial diagnostics
sp_diag(fit, sentencing)

# county `random effects' 
sp = spatial(fit)

# extract matrix of samples from posterior distribution of parameters
## alpha_re are the unstructured area random effects
S <- as.matrix(fit, pars = "alpha_re")

# extract data.frame of posterior samples
S <- as.data.frame(fit, pars = "alpha_re")
}