The Moran coefficient

The Moran coefficient, a measure of spatial autocorrelation (also known as Global Moran's I)

mc(x, w, digits = 3, warn = TRUE)

Arguments

x
Numeric vector of input values, length n.

w
An n x n spatial connectivity matrix. See shape2mat.

digits
Number of digits to round results to.

warn
If `FALSE`, no warning will be printed to inform you when observations with zero neighbors have been dropped.

Source

Chun, Yongwan, and Daniel A. Griffith. Spatial statistics and geostatistics: theory and applications for geographic information science and technology. Sage, 2013.

Cliff, Andrew David, and J. Keith Ord. Spatial processes: models & applications. Taylor & Francis, 1981.

Value

The Moran coefficient, a numeric value.

Details

If any observations with no neighbors are found (i.e. any(Matrix::rowSums(w) == 0)) they will be dropped automatically and a message will print stating how many were dropped.

See also

Examples

library(sf)
data(georgia)
w <- shape2mat(georgia, style = "W")
x <- georgia$ICE
mc(x, w)
#> [1] 0.519