Data model diagnostics

Visual diagnostics for spatial measurement error models.

me_diag(
  fit,
  varname,
  shape,
  w,
  probs = c(0.025, 0.975),
  plot = TRUE,
  size = 0.25,
  index = 0
)

Arguments

fit
A geostan_fit model object as returned from a call to one of the geostan::stan_* functions.

varname
Name of the modeled variable (a character string, as it appears in the model formula).

shape
An object of class sf or another spatial object coercible to sf with sf::st_as_sf such as SpatialPolygonsDataFrame.

w
An optional spatial connectivity matrix; if not provided, then a row-standardized adjacency matrix will be created using shape2mat(shape, "W").

probs
Lower and upper quantiles of the credible interval to plot.

plot
If FALSE, return a data.frame with the raw data values and posterior summary of the modeled variable.

size
Size of points and lines, passed to geom_pointrange.

index
Integer value; use this if you wish to identify observations with the largest `n=index` absolute Delta values; data on the top `n=index` observations ordered by absolute Delta value will be printed to the console and the plots will be labeled with the indices of the identified observations.

Value

A grid of spatial diagnostic plots for measurement error (i.e. data) models comparing the raw observations to the posterior distribution of the true values (given the specified observations, standard errors, and model). The Moran scatter plot and map depict the difference between the posterior means and the raw observations (i.e. shrinkage).

See also

Examples

if (FALSE) {
library(sf)
data(ohio)
C <- shape2mat(ohio)
ME <- list(se = data.frame(unemployment.acs = ohio$unemployment.acs.se),
           spatial = TRUE
         )
fit <- stan_glm(gop_growth ~ unemployment.acs,
                ME = ME,
                C = C,
                data = ohio, 
                prior_only = TRUE,
                refresh = 0
                )
me_diag(fit, "unemployment.acs", ohio)
me_diag(fit, "unemployment.acs", ohio, index = 3)
}