Data model diagnostics

Visual diagnostics for spatial measurement error models.

me_diag(
  fit,
  varname,
  shape,
  probs = c(0.025, 0.975),
  plot = TRUE,
  mc_style = c("scatter", "hist"),
  size = 0.25,
  index = 0,
  style = c("W", "B"),
  w = shape2mat(shape, match.arg(style)),
  binwidth = function(x) 0.5 * sd(x)
)

Source

Donegan, Connor and Chun, Yongwan and Griffith, Daniel A. (2021). ``Modeling community health with areal data: Bayesian inference with survey standard errors and spatial structure.'' Int. J. Env. Res. and Public Health 18 (13): 6856. DOI: 10.3390/ijerph18136856 Data and code: https://github.com/ConnorDonegan/survey-HBM.

Arguments

fit

A geostan_fit model object as returned from a call to one of the geostan::stan_* functions.

varname

Name of the modeled variable (a character string, as it appears in the model formula).

shape

An object of class sf or another spatial object coercible to sf with sf::st_as_sf.

probs

Lower and upper quantiles of the credible interval to plot.

plot

If FALSE, return a list of ggplots and a data.frame with the raw data values alongside a posterior summary of the modeled variable.

mc_style

Character string indicating how to plot the Moran coefficient for the delta values: if mc = "scatter", then moran_plot will be used with the marginal residuals; if mc = "hist", then a histogram of Moran coefficient values will be returned, where each plotted value represents the degree of residual autocorrelation in a draw from the join posterior distribution of delta values.

size

Size of points and lines, passed to geom_pointrange.

index

Integer value; use this if you wish to identify observations with the largest n=index absolute Delta values; data on the top n=index observations ordered by absolute Delta value will be printed to the console and the plots will be labeled with the indices of the identified observations.

style

Style of connectivity matrix; if w is not provided, style is passed to shape2mat and defaults to "W" for row-standardized.

w

An optional spatial connectivity matrix; if not provided, one will be created using shape2mat.

binwidth

A function with a single argument that will be passed to the binwidth argument in geom_histogram. The default is to set the width of bins to 0.5 * sd(x).

Value

A grid of spatial diagnostic plots for measurement error models comparing the raw observations to the posterior distribution of the true values. Includes a point-interval plot, a histogram of Moran coefficient values for the posterior distribution of Delta values (Delta = z - x, where z are the survey estimates and x are the modeled true values), and a map of the posterior mean of the Delta values.

See also

Examples

if (FALSE) {
library(sf)
data(georgia)
## binary adjacency matrix
A <- shape2mat(georgia, "B")
## prepare data for the CAR model, using WCAR specification
cp <- prep_car_data(A, type = "WCAR")
## provide list of data for the measurement error model
ME <- list(se = data.frame(ICE = georgia$ICE.se),
           spatial = TRUE,
           car_parts = cp
         )
## sample from the prior probability model only, including the ME model
fit <- stan_glm(log(rate.male) ~ ICE,
                ME = ME,
                C = C,
                data = georgia, 
                prior_only = TRUE,
                refresh = 0
                )
## see ME diagnostics
me_diag(fit, "ICE", georgia)
## see index values for the largest (absolute) delta values
 ## (differences between raw estimate and the posterior mean)
me_diag(fit, "ICE", georgia, index = 3)
}