Draw samples from the posterior predictive distribution

Draw samples from the posterior predictive distribution of a fitted geostan model. Use the original data or new data, with or without the spatial component and other partial pooling terms.

posterior_predict(
  object,
  newdata,
  C,
  samples,
  predictive = TRUE,
  re_form = NULL,
  spatial = TRUE,
  summary = FALSE,
  width = 0.95,
  seed,
  centerx = FALSE,
  scalex = FALSE
)

Arguments

object
A geostan_fit object.

newdata
A data.frame or matrix with data to use for predictions. Must be named and the names must match the model formula. If the model has an slx term and spatial = TRUE then newdata must have the same number of rows as the data used to fit the model.

C
Spatial connectivity matrix for calculating spatially lagged covariates (slx). Only used if spatial = TRUE and the model contains an slx term. It will be row-standardized automatically.

samples
Number of samples to take from the posterior distribution. Maximum is the total number of samples stored in the model.

predictive
Return the predictive distribution? Defaults to TRUE. If FALSE, then the linear predictor is returned after applying the inverse link function (results are on the same scale as the outcome variable).

re_form
If re_form = NA any random effects terms will be ignored when making predictions. If default re_form = NULL is set, then the results will include random effects terms. Those are the only options; this syntax is used to maintain some consistency with other R packages which have additional options.

spatial
Include the spatial component in the model predictions? Defaults to TRUE and will be ignored if the model does not have a spatial component (i.e. stan_glm). For models fit by stan_esf, stan_icar, and stan_car this option requires that newdata have the same number of observations as the data that the model was fit to.

summary
Should the posterior predictions be summarized using quantile intervals? It summary = TRUE the function returns a data frame with the expected value of the posterior distribution (column mu) and 100*width credible intervals for, respectively, the posterior predictive distribution (columns pred.lwr, pred.upr) and mean outcome (columns mu.lwr, mu.upr).

width
Only used if summary = TRUE to set the quantiles for the credible intervals.

seed
A single integer value to be used in a call to set.seed before taking samples from the posterior distribution. Passing a value to seed enables you to obtain the same results each time (by using the same seed).

centerx
Should newdata be centered using the means of the variables in the original data used to fit the model (stored in object$scale_params)? Defaults to FALSE.

scalex
Should newdata be centered and scaled using the standard deviations of the variables in the original data used to fit the model (stored in object$scale_params)? Defaults to FALSE.

Value

A matrix of size S x N containing samples from the posterior predictive distribution, where S is the number of samples and N is the number of observations. It is of class matrix and ppd.

If summary = TRUE the return value is a data frame with the expected value of the posterior distribution (column mu) and 100*width credible intervals for, respectively, the posterior predictive distribution (columns pred.lwr, pred.upr) and mean outcome (columns mu.lwr, mu.upr). The width of the intervals and the quantiles corresponding to the upper and lower limits of the credible intervals are returned as an attribute of the data frame (use attributes(df)$interval to retrieve them, where df is the data frame returned by posterior_predict).

Examples

if (FALSE) {
library(bayesplot)
library(ggplot2)
library(sf)
data(ohio)

## fit Ohio election model with a spatial filter
fit <- stan_esf(gop_growth ~ 1,
                data = ohio,
                C = shape2mat(ohio),
                refresh = 0
) 

# posterior predictive check (density overlay)
 # compare distribution of observations to the predictive distribution
y <- ohio$gop_growth
yrep <- posterior_predict(fit, samples = 75)
ppc_dens_overlay(y, yrep)

## plot marginal effects with credible intervals 
fit <- stan_esf(gop_growth ~ poly(college_educated, df = 3),
                data = ohio,
                C = shape2mat(ohio),
                refresh = 0
)

x <- seq(10, 45, by = 1)
newdata <- data.frame(college_educated = x)
pred.df <- posterior_predict(fit, newdata, spatial = FALSE, summary = TRUE)

## plot the mean of the predictive distribution (regression line)
a <- ggplot(pred.df, aes(x = mu)) +
    geom_line(aes( y = mu))

## add the credible interval for the mean
a <- a +
    geom_ribbon(aes(ymin = mu.lwr,
                    ymax = mu.upr),
               fill = "blue", alpha = 0.5)

## add the 95% credible interval for predictive distribution
a +
  geom_ribbon(aes(ymin = pred.lwr,
                  ymax = pred.upr),
              fill = "blue", alpha = 0.5)
}