Prepare data for a Stan CAR model

Prepare data for a Stan CAR model

prep_car_data(
  A,
  style = c("WCAR", "ACAR", "DCAR"),
  k = 1,
  gamma = 0,
  lambda = TRUE,
  cmat = TRUE
)

Arguments

A
Binary adjacency matrix; for style = DCAR, provide a symmetric matrix of distances instead. The distance matrix should be sparse, meaning that most distances should be zero (usually obtained by setting some threshold distance beyond which all are zero).

style
Specification for the connectivity matrix (C) and conditional variances (M); one of "WCAR", "ACAR", or "DCAR".

k
For style = DCAR, distances will be raised to the -k power (d^-k).

gamma
For style = DCAR, distances will be offset by gamma before raising to the -kth power.

lambda
If TRUE, return eigenvalues required for calculating the log determinant of the precision matrix and for determining the range of permissible values of rho. These will also be printed with a message if lambda = TRUE.

cmat
If cmat = TRUE, return the full matrix C (in sparse matrix format).

Source

Cliff A, Ord J (1981). Spatial Processes: Models and Applications. Pion.

Cressie N (2015 1993). Statistics for Spatial Data. Revised edition. John Wiley & Sons.

Cressie N, Perrin O, Thomas-Agnan C (2005). “Likelihood-based estimation for Gaussian MRFs.” Statistical Methodology, 2(1), 1–16.

Cressie N, Wikle CK (2011). Statistics for Spatio-Temporal Data. John Wiley & Sons.

Haining RP, Li G (2020). Modelling Spatial and Spatio-Temporal Data: A Bayesian Approach. CRC Press.

Value

A list containing all of the data elements required by the Stan CAR model.

Details

The CAR model is:

  Normal(Mu, Sigma), Sigma = (I - rho * C)^-1 * M * tau^2,

where I is the identity matrix, rho is a spatial autocorrelation parameter, C is a connectivity matrix, and M * tau^2 is a diagonal matrix with conditional variances on the diagonal. tau^2 is a (scalar) scale parameter.

In the WCAR specification, C is the row-standardized version of A. This means that the non-zero elements of A will be converted to 1/N_i where N_i is the number of neighbors for the ith site (obtained using Matrix::rowSums(A). The conditional variances (on the diagonal of M * tau^2), are also proportional to 1/N_i.

The ACAR specification is from Cressie, Perrin and Thomas-Agnon (2005); also see Cressie and Wikle (2011, p. 188).

The DCAR specification is inverse distance-based, and requires the user provide a (sparse) distance matrix instead of a binary adjacency matrix. (For A, provide a symmetric matrix of distances, not inverse distances!) Internally, non-zero elements of A will be converted to: d_{ij} = (a_{ij} + gamma)^(-k) (Cliff and Ord 1981, p. 144). Default values are k=1 and gamma=0. Following Cressie (1993), these values will be standardized by the maximum d_{ij} value. The conditional variances will be proportional to the inverse of the row sums of the transformed distance matrix: D_{ii} = (sum_i^N d_{ij})^(-1).

For inverse-distance weighting schemes, see Cliff and Ord (1981); for distance-based CAR specifications, see Cressie (2015 1993) and Haining and Li (2020).

When using stan_car, always use cmat = TRUE (the default).

Examples

data(georgia)

## binary adjacency matrix
A <- shape2mat(georgia, style = "B")

## get list of data for Stan
cp <- prep_car_data(A, "WCAR")

#> Range of permissible rho values:  -1.661134 1 

 # range of permissible values for rho:
1 / range(cp$lambda)

#> [1] -1.661134  1.000000

if (FALSE) {
## pass the data to stan_car
fit = stan_car(log(rate.male) ~ 1, data = georgia, car_parts = cp)
}

# or use the ACAR specification
cp <- prep_car_data(A, "ACAR")

#> Range of permissible rho values:  -0.3283182 0.1634063 

 # range of permissible values for rho:
1 / range(cp$lambda)

#> [1] -0.3283182  0.1634063