Create spatial and space-time connectivity matrices

A wrapper function for a string of spdep (and other) functions required to convert spatail objects to spatial or spatio-temporal connectivity matrices.

shape2mat(
  shape,
  style = c("B", "W", "C", "S"),
  t = 1,
  st.type = "contemp",
  zero.policy = TRUE,
  queen = TRUE,
  snap = sqrt(.Machine$double.eps)
)

Arguments

shape
An object of class sf, SpatialPolygons or SpatialPolygonsDataFrame.

style
What kind of coding scheme should be used to create the spatial connectivity matrix? Defaults to "B" for binary; use "W" for row-standardized weights; "C" for globally standardized and "S" for the Tiefelsdorf et al.'s (1999) variance-stabilizing scheme. This is passed internally to nb2mat.

t
Number of time periods. Currently only the binary coding scheme is available for space-time connectivity matrices.

st.type
For space-time data, what type of space-time connectivity structure should be used? Options are "lag" for the lagged specification and "contemp" (the default) for contemporaneous specification.

zero.policy
Are regions with zero neighbors allowed? Default zero.policy = TRUE (allowing regions to have zero neighbors). Also passed to nb2mat.

queen
Passed to poly2nb to set the contiguity condition. Defaults to TRUE so that a single shared boundary point between polygons is sufficient for them to be considered neighbors.

snap
Passed to poly2nb; "boundary points less than ‘snap’ distance apart are considered to indicate contiguity."

Source

Griffith, D. A. (2012). Space, time, and space-time eigenvector filter specifications that account for autocorrelation. Estadística Espanola, 54(177), 7-34.

Griffith, D. A., Chun, Y., Li, B. (2020). Spatial Regression Analysis Using Eigenvector Spatial Filtering. Academic Press, Ch. 8.

Haining, R. P., & Li, G. (2020). Regression Modelling Wih Spatial and Spatial-Temporal Data: A Bayesian Approach. CRC Press.

Value

A spatial connectivity matrix

Details

Haining and Li (Ch. 4) provide a helpful discussion of spatial connectivity matricies (Ch. 4) and spatio-temporal connectivity matricies (Ch. 15).

The `lagged' space-time structure connects each observation to its own past (one period lagged) value and the past value of its neighbors. The `contemporaneous' specification links each observation to its neighbors and to its own in situ past (one period lagged) value (Griffith 2012, p. 23).

See also

Examples

data(ohio)
C <- shape2mat(ohio, "B")
W <- shape2mat(ohio, "W")

## for space-time data
## if you have multiple years with same neighbors
## provide the geography (for a single year!) and number of years \code{t}
## defaults to the contemporaneous connectivity structure
Cst <- shape2mat(ohio, t = 5)