Create spatial and space-time connectivity matrices

Creates sparse matrix representations of spatial connectivity structures

shape2mat(
  shape,
  style = c("B", "W"),
  queen = TRUE,
  snap = sqrt(.Machine$double.eps),
  t = 1,
  st.style = c("contemp", "lag")
)

Source

Bivand, Roger S. and Pebesma, Edzer and Gomez-Rubio, Virgilio (2013). Applied spatial data analysis with R, Second edition. Springer, NY. https://asdar-book.org/

Griffith, Daniel A. (2012). Space, time, and space-time eigenvector filter specifications that account for autocorrelation. Estadística Espanola, 54(177), 7-34.

Haining, Robert P. and Li, Guangquan (2020). Regression Modelling Wih Spatial and Spatial-Temporal Data: A Bayesian Approach. CRC Press.

Arguments

shape

An object of class sf, SpatialPolygons or SpatialPolygonsDataFrame.

style

What kind of coding scheme should be used to create the spatial connectivity matrix? Defaults to "B" for binary; use "W" for row-standardized weights.

queen

Passed to poly2nb to set the contiguity condition. Defaults to TRUE so that a single shared boundary point (rather than a shared border/line) between polygons is sufficient for them to be considered neighbors.

snap

Passed to poly2nb; "boundary points less than ‘snap’ distance apart are considered to indicate contiguity."

t

Number of time periods. Only the binary coding scheme is available for space-time connectivity matrices.

st.style

For space-time data, what type of space-time connectivity structure should be used? Options are "lag" for the lagged specification and "contemp" (the default) for contemporaneous specification (see Details).

Value

A spatial connectivity matrix

Details

Haining and Li (Ch. 4) provide a helpful discussion of spatial connectivity matrices (Ch. 4).

The space-time connectivity matrix can be used for eigenvector space-time filtering (stan_esf. The lagged' space-time structure connects each observation to its own past (one period lagged) value and the past value of its neighbors. The contemporaneous' specification links each observation to its neighbors and to its own in situ past (one period lagged) value (Griffith 2012, p. 23).

See also

Examples

data(georgia)

## binary adjacency matrix
C <- shape2mat(georgia, "B")
## row sums gives the numbers of neighbors per observation
Matrix::rowSums(C)
#>   [1]  6  5  6  5  3  5  7  4  5  5  5  7  6  4  5  4  7  7  4  5  7  6  6  6  4
#>  [26]  5  6  3  8  6  5  7  3  3  8  6  6  7  5  8  8  5  6  6  5  4  5  6  7  5
#>  [51]  2  7  6  5  3  1  3  7  6  2  5  5  8  6  7  7  5  6  4  4  6  4  6 10  8
#>  [76]  7  7  6  6  3  6 10  6  6  5  2  6  5  4  7  4  8  5  4  5  4  4  7  6  6
#> [101]  6  5  4  4  6  8  4  6  6  4  5  4  4  4  9  3  4  7  7  4  3  7  4  6  4
#> [126]  5  7  6  4  4  5  6  5  5  6  6  7  5  6  5  4  7  6  5  7  3  3  7  3  7
#> [151]  5  6  5  4  6  3  5  7  8
head(Matrix::summary(C))
#> 159 x 159 sparse Matrix of class "ngCMatrix", with 860 entries 
#>     i j
#> 1  23 1
#> 2  58 1
#> 3  59 1
#> 4 131 1
#> 5 148 1
#> 6 159 1

## row-standardized matrix 
W <- shape2mat(georgia, "W")
Matrix::rowSums(W)
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 1 1 1 1 1 1 1 1 1 1
head(Matrix::summary(W))
#> 159 x 159 sparse Matrix of class "dgCMatrix", with 860 entries 
#>     i j         x
#> 1  23 1 0.1666667
#> 2  58 1 0.1428571
#> 3  59 1 0.1666667
#> 4 131 1 0.2000000
#> 5 148 1 0.1428571
#> 6 159 1 0.1250000

## space-time matricies 
## for eigenvector space-time filtering
## if you have multiple years with same neighbors,
## provide the geography (for a single year!) and number of years \code{t}
Cst <- shape2mat(georgia, t = 5)
dim(Cst)
#> [1] 795 795
EVst <- make_EV(Cst)
dim(EVst)
#> [1] 795 208