Conditional autoregressive models

Fit a regression model with a conditional auto-regressive (CAR) spatial component. Only binary weighting schemes are currently supported.

stan_car(
  formula,
  slx,
  re,
  data,
  ME = NULL,
  C,
  EV,
  family = poisson(),
  prior = NULL,
  prior_intercept = NULL,
  prior_tau = NULL,
  prior_phi_precision = c(2, 2),
  centerx = FALSE,
  scalex = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 10000,
  refresh = 500,
  pars = NULL,
  control = list(adapt_delta = 0.9, max_treedepth = 15),
  silent = FALSE,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models can be specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally). These will be pre-multiplied by a row-standardized version of the user-provided spatial weights matrix and then added (prepended) to the design matrix. If and when setting priors for beta manually, remember to include priors for any SLX terms as well.

re
If the model includes a varying intercept term (or "spatially unstructured random effect") specify the grouping variable here using formula syntax, as in ~ ID.

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

ME
To model observational error (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. Errors are assigned a Gaussian probability distribution and the modeled (true) covariate vector is assigned a Student's t model with optional spatially varying mean. Elements of the list ME may include:

se

a dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounded

If any variables in se are bounded within some range (e.g. percentages ranging from zero to one hundred) provide a vector of zeros and ones indicating which columns are bounded. By default the lower bound will be 0 and the upper bound 100, for percentages.

bounds

A numeric vector of length two providing the upper and lower bounds, respectively, of the bounded variables. Defaults to bounds = c(0, 100).

spatial

Logical value indicating if the models for covariates should include a spatially varying mean (using an eigenvector spatial filter). Defaults to spatial = FALSE. If spatial = TRUE and you do not provide both ME$prior_rhs and EV then you must provide a connectivity matrix C.

prior_rhs

Optional prior parameters for the regularized horseshoe (RHS) prior used for the ESF data model; only used if ME$spatial = TRUE. The RHS prior is used for the eigenvector spatial filter (ESF), as in stan_esf. Must be a named list containing vectors slab_df, slab_scale, scale_global, and varname. The character vector varname indicates the order of the other parameters (by name).

C
Binary spatial connectivity matrix. This is used for the CAR prior and to calculate residual spatial autocorrelation as well as any user specified slx terms (it will be row-standardized before calculating slx terms).

EV
A matrix of eigenvectors from any (transformed) connectivity matrix (presumably spatial). See make_EV and shape2mat.

family
The likelihood function for the outcome variable. Current options are binomial(link = "logit") and poisson(link = "log").

prior
A data.frame or matrix with location and scale parameters for Gaussian prior distributions on the model coefficients. Provide two columns---location and scale---and a row for each variable in their order of appearance in the model formula. Default priors are weakly informative relative to the scale of the data.

prior_intercept
A vector with location and scale parameters for a Gaussian prior distribution on the intercept; e.g. prior_intercept = c(0, 10).

prior_tau
Set hyperparameters for the scale parameter of varying intercepts re. The varying intercepts are given a normal prior with scale parameter alpha_tau. The latter is given a half-Student's t prior with default of 20 degrees of freedom, centered on zero and scaled to the data to be weakly informative. To adjust it use, e.g., prior_tau = c(df = 15, location = 0, scale = 5).

prior_phi_precision
Hyperprior parameters for the precision (inverse scale) of the CAR component phi, a numeric vector of length two. The precision parameter for phi is phi_tau which is assigned a Gamma prior distribution; defaults to prior_phi_precision = c(2, 2).

centerx
Logical value indicating if the covariates should be centered prior to fitting the model.

scalex
Logical value indicating if the covariates be centered and scaled (divided by their standard deviation).

prior_only
Logical value; if TRUE, draw samples only from the prior distributions of parameters.

chains
Number of MCMC chains to estimate.

iter
Number of samples per chain.

refresh
Stan will print the progress of the sampler every refresh number of samples. Set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model.

control
A named list of parameters to control the sampler's behavior. See stan for details. The defaults are the same rstan::stan except that adapt_delta is raised to .9 and max_treedepth = 15.

silent
If TRUE, suppress printed messages including prior specifications and Stan sampling progress (i.e. refresh=0). Stan's error and warning messages will still print.

...
Other arguments passed to sampling. For multi-core processing, you can use cores = parallel::detectCores(), or run options(mc.cores = parallel::detectCores()) first.

Source

Joseph, Max (2016). Exact Sparse CAR Models in Stan. Stan Case Studies, Vol. 3. https://mc-stan.org/users/documentation/case-studies/mbjoseph-CARStan.html

Lawson, Andrew B. (2013). Bayesian Disease Mapping: Hierarchical Modeling in Spatial Epidemiology. CRC Press.

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame, including phi the spatial term, phi_tau the scale of the CAR model, and phi_alpha the parameter that controls the degree of spatial autocorrelation in phi and has range [0, 1].

diagnostic

Widely Applicable Information Criteria (WAIC) with crude measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and residual spatial autocorrelation (Moran coefficient of the residuals). Residuals are relative to the mean posterior fitted values.

stanfit

an object of class stanfit returned by rstan::stan

data

a data frame containing the model data

edges

The edge list representing all unique sets of neighbors (a sparse representation of C)

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including CAR component)

slx

The slx formula

re

A list containing re, the random effects (varying intercepts) formula if provided, and Data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

scale_params

A list with the center and scale parameters returned from the call to base::scale on the model matrix. If centerx = FALSE and scalex = FALSE then it is an empty list.

spatial

A data frame with the name of the spatial component parameter ("phi") and method ("CAR")

Details

The CAR model is the prior distribution for the parameter vector phi and has two associated parameters: phi_alpha which controls the degree of spatial autocorrelation (and thus the amount of spatial smoothing) and phi_tau which is a precision (inverse scale) parameter. The Stan code for the CAR component of the model is from Joseph (2016).

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(bayesplot)
library(sf)
options(mc.cores = parallel::detectCores())
data(sentencing)

C <- shape2mat(sentencing)
log_e <- log(sentencing$expected_sents)
fit.car <- stan_car(sents ~ offset(log_e),
                    family = poisson(),
                    data = sentencing,
                    C = C,
                    refresh = 0
)

# posterior predictive check: predicted distribution should resemble observed distribution
yrep <- posterior_predict(fit.car, samples = 75)
y <- sentencing$sents
ppc_dens_overlay(y, yrep)

sp.trend <- spatial(fit.car)$mean
st_as_sf(sentencing) %>%
  ggplot() +
  geom_sf(aes(fill = sp.trend)) +
  scale_fill_gradient2()
 }