Conditional autoregressive (CAR) models

Use the CAR model as a prior on parameters, or fit data to an auto-Gaussian CAR model.

stan_car(
  formula,
  slx,
  re,
  data,
  car_parts,
  ME = NULL,
  family = gaussian(),
  prior = NULL,
  centerx = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 2000,
  refresh = 500,
  pars = NULL,
  control = NULL,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models can be specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally).

These will be pre-multiplied by a row-standardized version of the user-provided spatial weights matrix and then added (prepended) to the design matrix. For example, providing

stan_car(y ~ x1 + x2, slx = ~ x1, ...)

is the same as providing

stan_car(y ~ I(W %*% x1) + x1 + x2, ...)

where W is a row-standardized spatial weights matrix (see shape2mat). When setting priors for beta, remember to include priors for any SLX terms as well.

re
To include a varying intercept (or "random effects") term, alpha_re, specify the grouping variable here using formula syntax, as in ~ ID. Then, alpha_re is a vector of parameters added to the linear predictor of the model, and:

       alpha_re ~ N(0, alpha_tau)
       alpha_tau ~ Student_t(d.f., location, scale).

With the CAR model, any alpha_re term should be at a different level or scale than the observations; that is, at a different scale than the autocorrelation structure of the CAR model itself.

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

car_parts
A list of data for the CAR model, as returned by prep_car_data.

ME
To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. The ME models are designed for American Community Survey (ACS) estimates, x, and their standard errors, se (Donegan, Chun and Griffith 2021). The ME models have one of the the following two specifications, depending on the user input:

       x ~ Gauss(x_true, se)
       x_true ~ MVGauss(mu, Sigma)
       Sigma = (I - rho * C)^(-1) M * tau^2
       mu ~ Gauss(0, 100)
       tau ~ student_t(10, 0, 40)
       rho ~ uniform(lower_bound, upper_bound)

where the covariance matrix, Sigma, has the conditional autoregressive specification, and tau is the scale parameter. If ME$car_parts = FALSE, then a non-spatial model will be used instead:

       x ~ Gauss(x_true, se)
       x_true ~ student_t(df, mu, sigma)
       df ~ gamma(3, 0.2)
       mu ~ Gauss(0, 100)
       sigma ~ student_t(10, 0, 40)

The target of inference is x_true, the actual values of the covariate during the period of the survey, and the observational error is the difference between the survey estimate and the unknown value that would have been obtained by a complete census. Elements of the list ME may include:

se

A required dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounds

An optional numeric vector of length two providing the upper and lower bounds, respectively, of the variables. If not provided, they will be set to c(-Inf, Inf) (i.e., unbounded). Common usages include keeping percentages between zero and one hundred or proportions between zero and one.

prior

Optionally provide parameter values for the prior distributions of the measurement error model(s). The ME models contain a location parameter (mu) and a scale parameter (tau or sigma), each of which require prior distributions. If none are provided, default priors will be assigned and printed to the console.

The prior for the location parameter, mu, is Gaussian (the default being Gauss(0, 100)). You can alter this by providing a data.frame with columns named location and scale; provide values for each covariate in se. List the values in the same order as the columns of se.

The prior for the scale parameters is Student's t, and the default is Student_t(10, 0, 40). You can alter this by providing a data.frame with columns named df, location and scale; provide values for each covariate in se. List the values in the same order as the columns of se.

For example, if you are modeling two covariates with the ME models, you can add the priors to an existing ME list like this:
ME$prior <- list(location = data.frame(location = c(0, 0),
                                          scale = c(10, 10)),
                 scale    = data.frame(df  c(10, 10),
                                       location = c(0, 0),
                                       scale = c(10, 10)), 

Note that if a prior is provided, you must provide priors for both location (mu) and scale (tau or sigma).

The CAR model also has a spatial autocorrelation parameter, rho, which is assigned a uniform prior distribution. You can set the boundaries of the prior with:
ME$prior$car_rho <- c(lower_bound, upper_bound)

You must specify values that are within the permissible range of values for rho. This range is automatically printed to the console by prep_car_data.

car_parts

By default, any ME models within stan_car will be spatial CAR models. Any data passed to ME$car_parts will be ignored. However, to use a non-spatial Student's t ME model instead, provide ME$car_parts = FALSE.

family
The likelihood function for the outcome variable. Current options are auto_gaussian(), binomial(link = "logit"), and poisson(link = "log"); if family = gaussian() is provided, it will automatically be converted to auto_gaussian().

prior
A named list of parameters for prior distributions. If not provided, the default prior distributions will be assigned and printed to the console. User-defined priors can be assigned to the following parameters:

intercept

The intercept is assigned a Gaussian prior distribution; provide a numeric vector with location and scale parameters; e.g. prior = list(intercept = c(location = 0, scale = 10)) to assign a Gaussian prior with mean zero and variance 10^2.

beta

Regression coefficients are assigned Gaussian prior distributions. Provide a data.frame with two columns (location and scale).The order of variables must follow their order of appearance in the model formula; e.g., if the formula is y ~ x1 + x2, then providing
prior = list(beta = data.frame(location = c(0, 2),
                                  scale = c(1, 2)))

will assign the following prior distributions:
       x1 ~ Gauss(0, 1)
       x2 ~ Gauss(2, 2)

Note that if you also use slx terms (spatially lagged covariates), then you have to provide priors for them. If your model specification is:
 stan_glm(y ~ x1 + x2, slx = ~ x1, ...)

then the prior for beta must have three rows. Since slx terms are always prepended to the design matrix, the prior for the slx term will be listed first.

tau

The scale parameter for (spatially unstructured) random effects, or varying intercepts, terms. This scale parameter, tau, is assigned a half-Student's t prior. To set this, use, e.g., prior = list(tau = c(df = 20, location = 0, scale = 20)).

car_scale

Hyperprior parameters for the scale of the CAR model, car_scale. The scale is assigned a Student's t prior model; to set its parameter values, provide a length-three vector with the degrees of freedom, location, and scale parameters. E.g., prior = list(car_scale = c(df = 15, location = 0, scale = 2)).

car_rho

The spatial autocorrelation parameter in the CAR model, rho, is assigned a uniform prior distribution. By default, the prior will be uniform over all permissible values as determined by the eigenvalues of the connectivity matrix, C. You may alter those limits by providing a numeric vector such as: prior$car_rho = c(0, 1).The range of permissible values for rho is automatically printed to the R console by prep_car_data. You may also identify the range of permissible values using the eigenvalues, lambda, returned by prep_car_data and the following code: rho_limits = 1 / range(lambda). These limits are sensitive to the chosen specification (WCAR, ACAR, etc.) and connectivity matrix.

centerx
To center predictors on their mean values, use centerx = TRUE. If centerx is a numeric vector, predictors will be centered on the values provided. This argument is passed to scale.

prior_only
Logical value; if TRUE, draw samples only from the prior distributions of parameters.

chains
Number of MCMC chains to use.

iter
Number of samples per chain.

refresh
Stan will print the progress of the sampler every refresh number of samples. Set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model.

control
A named list of parameters to control the sampler's behavior. See stan for details.

...
Other arguments passed to sampling. For multi-core processing, you can use cores = parallel::detectCores(), or run options(mc.cores = parallel::detectCores()) first.

Source

Cressie, Noel (2015 (1993)). Statistics for Spatial Data. Wiley Classics, Revised Edition.

Cressie, Noel and Wikle, Christopher (2011). Statistics for Spatio-Temporal Data. Wiley.

Donegan, Connor and Chun, Yongwan and Griffith, Daniel A. (2021). Modeling community health with areal data: Bayesian inference with survey standard errors and spatial structure. Int. J. Env. Res. and Public Health 18 (13): 6856. DOI: 10.3390/ijerph18136856 Data and code: https://github.com/ConnorDonegan/survey-HBM.

Haining, Robert and Li, Guangquan (2020). Modelling Spatial and Spatial-Temporal Data: A Bayesian Approach. CRC Press.

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame.

diagnostic

Widely Applicable Information Criteria (WAIC) with a measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and mean residual spatial autocorrelation as measured by the Moran coefficient.

stanfit

an object of class stanfit returned by rstan::stan

data

a data frame containing the model data

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including CAR component)

slx

The slx formula

re

A list containing re, the varying intercepts (re) formula if provided, and Data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

x_center

If covariates are centered internally (i.e., centerx is not FALSE), then x_centers is the numeric vector of values on which the covariates were centered.

spatial

A data frame with the name of the spatial component parameter (either "phi" or, for auto Gaussian models, "trend") and method ("CAR")

ME

A list indicating if the object contains an ME model; if so, the user-provided ME list is also stored here.

C

Spatial connectivity matrix (in sparse matrix format).

Details

CAR models are discussed in Cressie and Wikle (2011, p. 184-88), Cressie (2015, Ch. 6-7), and Haining and Li (2020, p. 249-51).

The Stan code for this implementation of the CAR model first introduced in Donegan et al. (2021, supplementary material) for models of small area survey data.

Details and results depend on the family argument, as well as on the particular CAR specification chosen (see prep_car_data).

family = auto_gaussian()

With family = auto_gaussian() (referred to here as the auto-Gaussian model), the CAR model is specified as follows:

                            Y ~ MVGauss(Mu, Sigma)
                            Sigma = (I - rho C)^-1 * M * tau^2

where Mu is the mean vector (with intercept, covariates, etc.), C is a spatial connectivity matrix, and M is a known diagonal matrix with diagonal entries proportional to the conditional variances. C and M are provided by prep_car_data.

The covariance matrix of the CAR model, Sigma, contains two parameters: car_rho (rho), which controls the degree of spatial autocorrelation, and the scale parameter, car_scale (tau). The range of permissible values for rho depends on the specification of C and M; for options, see prep_car_data and Cressie and Wikle (2011, pp. 184-188).

The auto-Gaussian model contains an implicit spatial trend (i.e., autocorrelation) component which is calculated as follows (Cressie 2015, p. 564):

                            trend = rho * C * (Y - Mu).

This term can be extracted from a fitted auto-Gaussian model using the spatial method.

When applied to a fitted auto-Gaussian model, the residuals.geostan_fit method returns `de-trended' residuals by default. That is,

                            residual = Y - Mu - trend.

To obtain "raw" residuals (Y - Mu), use residuals(fit, detrend = FALSE).

family = poisson()

For family = poisson(), the model is specified as:

                            Y ~ Poisson(exp(offset + lambda))
                            lambda ~ MVGauss(Mu, Sigma)
                            Sigma = (I - rho C)^-1 * M * tau^2

These models are most often used to calculate small area incidence rates (mortality or disease incidence rates); the user provided offset should be, then, the natural logarithm of the denominator in the rates, e.g., log-population at risk.

For Poisson models, the spatial method returns the parameter vector phi, which is the log-risk minus the intercept and any covariates:

                           phi = lambda - Mu.

This is the spatial autocorrelation component. This is equivalent to specifying the model as:

                            Y ~ Poisson(exp(offset + Mu + phi))
                            phi ~ MVGauss(0, Sigma)
                            Sigma = (I - rho C)^-1 * M * tau^2.

In the Poisson CAR model, phi contains a latent spatial trend as well as additional variation around it. If you would like to extract the latent/implicit spatial trend from phi, you can do so by calculating (following Cressie 2015, p. 564):

                            trend = rho * C * phi.

family = binomial()

For family = binomial(), the model is specified as:

                            Y ~ Binomial(N, theta)
                            logit(theta) ~ MVGauss(Mu, Sigma)
                            Sigma = (I - rho C)^-1 * M * tau^2

where outcome data Y are counts, N is the number of trials, and theta is the 'success' rate. Note that the model formula should be structured as: cbind(sucesses, failures) ~ x, such that trials = successes = failures.

For fitted Binomial models, the spatial method will return the parameter vector phi, equivalent to:

                            phi = logit(theta) - Mu.

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(bayesplot)
library(sf)
options(mc.cores = parallel::detectCores())
data(sentencing)

C <- shape2mat(sentencing, style = "B")
car.dl <- prep_car_data(C, style = "WCAR")
log_e <- log(sentencing$expected_sents)
fit.car <- stan_car(sents ~ offset(log_e),
                    family = poisson(),
                    data = sentencing,
                    car_parts = car.dl
)

# MCMC diagnostics
rstan::stan_rhat(fit.car$stanfit)
rstan::stan_ess(fit.car$stanfit)

# Spatial diagnostics
sp_diag(fit.car, sentencing)

# posterior predictive distribution
yrep <- posterior_predict(fit.car, S = 75)
y <- sentencing$sents
ppc_dens_overlay(y, yrep)

# examine posterior distributions of CAR parameters
plot(fit.car, pars = c("car_scale", "car_rho"))

# map the spatial autocorrelation term, phi
sp.trend <- spatial(fit.car)$mean
st_as_sf(sentencing) %>%
  ggplot() +
  geom_sf(aes(fill = sp.trend)) +
  scale_fill_gradient2()
 
# calculate log-standardized sentencing ratios (log-SSRs)
# (like Standardized Incidence Ratios: observed/exected case counts)
SSR <- fitted(fit.car)$mean
log.SSR <- log( SSR, base = 2 )

st_as_sf(sentencing) %>%
 ggplot() +
 geom_sf(aes(fill = log.SSR)) +
 scale_fill_gradient2(
   midpoint = 0,
   name = NULL,
   breaks = seq(-3, 3, by = 0.5)
   ) +
 labs(title = "Log-Standardized Sentencing Ratios",
      subtitle = "log( Fitted/Expected ), base 2"
      ) +
 theme_void() +
 theme(
   legend.position = "bottom",
   legend.key.height = unit(0.35, "cm"),
   legend.key.width = unit(1.5, "cm")
 )
}