Conditional autoregressive (CAR) models

Apply the CAR model as a prior on parameters, or fit an auto Gaussian model.

stan_car(
  formula,
  slx,
  re,
  data,
  ME = NULL,
  car_parts,
  family = gaussian(),
  invert = TRUE,
  prior = NULL,
  prior_intercept = NULL,
  prior_tau = NULL,
  prior_car_scale = NULL,
  centerx = FALSE,
  scalex = FALSE,
  prior_only = FALSE,
  chains = 5,
  iter = 2000,
  refresh = 500,
  pars = NULL,
  control = list(adapt_delta = 0.9, max_treedepth = 13),
  silent = FALSE,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models can be specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally). These will be pre-multiplied by a row-standardized version of the user-provided spatial weights matrix and then added (prepended) to the design matrix. If and when setting priors for beta manually, remember to include priors for any SLX terms as well.

re
If the model includes a varying intercept term alpha_re specify the grouping variable here using formula syntax, as in ~ ID. Then, alpha_re ~ N(0, alpha_tau), alpha_tau ~ Student_t(d.f., location, scale). With the CAR model, any alpha_re term should be at a different level or scale than the observations (i.e., a different scale than the autocorrelation structure of the CAR model itself).

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

ME
To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. This implements the methodology introduced by Donegan et al. (2021). Errors are assigned a Gaussian probability distribution and the modeled (true) covariate vector is assigned an auto-Gaussian (CAR) model unless spatial = FALSE, in which case they are assinged the Student's t model. Elements of the list ME may include:

se

a dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounded

If any variables in se are bounded within some range (e.g. percentages ranging from zero to one hundred) provide a vector of zeros and ones indicating which columns are bounded. By default, if bounded = TRUE, the lower bound will be 0 and the upper bound 100, for percentages.

bounds

A numeric vector of length two providing the upper and lower bounds, respectively, of any bounded variables.

spatial

Logical value indicating whether an auto-Gaussian (conditional autoregressive (CAR)) model should be used for the covariates. For stan_car, defaults to spatial = TRUE.

car_parts

This argument is ignored for stan_car, because the user must always provide this information within the main function call. See the car_parts argument below. If any information is provided to ME$car_parts, it will be overwritten internally.

car_parts
A list of data for the CAR model, as returned by prep_car_data (be sure to return the matrix C, by using the argument cmat = TRUE).

family
The likelihood function for the outcome variable. Current options are gaussian(), binomial(link = "logit"), and poisson(link = "log").

invert
To calculate the log likelihood of the data log_lik and the posterior predictive distribution yrep with the auto-Gaussian model, the precision matrix needs to be inverted. This can be costly for large data sets---the inversion needs to be completed once per posterior sample. To avoid the computational cost, set invert = FALSE. Note, this is only used when family = gaussian().

prior
A data.frame or matrix with location and scale parameters for Gaussian prior distributions on the regression coefficients. Provide two columns---location and scale---and a row for each variable in their order of appearance in the model formula. Default priors are weakly informative relative to the scale of the data.

prior_intercept
A vector with location and scale parameters for a Gaussian prior distribution on the intercept; e.g. prior_intercept = c(0, 10).

prior_tau
Set hyperparameters for the scale parameter of varying intercepts alpha_re. The varying intercepts are given a normal prior with scale parameter alpha_tau. alpha_tau is given a half-Student's t prior with default of 20 degrees of freedom, centered on zero and scaled to the data to be weakly informative. To adjust it use, e.g., prior_tau = c(df = 15, location = 0, scale = 5).

prior_car_scale
Hyperprior parameters for the scale of the CAR model car_scale. The scale is assigned a Student's t prior model; to set its parameter values provide a length-three vector with the degrees of freedom, location, and scale parameters. E.g., prior_car_scale = c(df = 15, location = 0, scale = 2).

centerx
Logical value indicating if the covariates should be centered prior to fitting the model.

scalex
Logical value indicating if the covariates be centered and scaled (divided by their standard deviation).

prior_only
Logical value; if TRUE, draw samples only from the prior distributions of parameters.

chains
Number of MCMC chains to use.

iter
Number of samples per chain.

refresh
Stan will print the progress of the sampler every refresh number of samples. Set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model.

control
A named list of parameters to control the sampler's behavior. See stan for details. The defaults are the same as stan except that adapt_delta is raised to .9 and max_treedepth = 15.

silent
If TRUE, suppress printed messages including prior specifications and Stan sampling progress (i.e. refresh=0). Stan's error and warning messages will still print.

...
Other arguments passed to sampling. For multi-core processing, you can use cores = parallel::detectCores(), or run options(mc.cores = parallel::detectCores()) first.

Source

Cressie, Noel (2015 1993). Statistics for Spatial Data. Wiley Classics, Revised Edition.

Cressie, Noel and Wikle, Christopher (2011). Statistics for Spatio-Temporal Data. Wiley.

Donegan, Connor and Chun, Yongwan and Griffith, Daniel A. (2021). ``Modeling community health with areal data: Bayesian inference with survey standard errors and spatial structure.'' Int. J. Env. Res. and Public Health 18 (13): 6856. DOI: 10.3390/ijerph18136856 Data and code: https://github.com/ConnorDonegan/survey-HBM.

Haining, Robert and Li, Guangquan (2020). Modelling Spatial and Spatial-Temporal Data: A Bayesian Approach. CRC Press.

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame.

diagnostic

Widely Applicable Information Criteria (WAIC) with crude measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and residual spatial autocorrelation (Moran coefficient of the residuals). Residuals are relative to the mean posterior fitted values.

stanfit

an object of class stanfit returned by rstan::stan

data

a data frame containing the model data

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including CAR component)

slx

The slx formula

re

A list containing re, the varying intercepts (re) formula if provided, and Data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

scale_params

A list with the center and scale parameters returned from the call to base::scale on the model matrix. If centerx = FALSE and scalex = FALSE then it is an empty list.

spatial

A data frame with the name of the spatial component parameter (either "phi" or, for auto Gaussian models, "trend") and method ("CAR")

Details

CAR models are discussed in Cressie and Wikle (2011, p. 184-88), Cressie (2015 1993, Ch. 6-7), and Haining and Li (2020, p. 249-51).

The Stan code for this implementation of the CAR model first introduced in Donegan et al. (2021, supplementary material) for models of small area survey data.

Details and results depend on the family argument (see also prep_car_data).

family = gaussian()

With family = gaussian() (referred to here as the auto-Gaussian model), the CAR model is specified as follows:

                            Y ~ MVGauss(Mu, Sigma)

                            Sigma = (I - rho C)^-1 * M * tau^2

where Mu is the mean vector (with intercept, covariates, etc.), C is the spatial connectivity matrix, and M is a known diagonal matrix with diagonal entries proportional to the conditional variances (see the M_diagonal argument).

The covariance matrix of the CAR model, Sigma, contains two parameters to estimate: car_rho (rho), which controls the degree of spatial autocorrelation, and the scale parameter, car_scale (tau). The range of permissible values for rho depends on the specification of C and M; for options, see prep_car_data and Cressie and Wikle (2011, pp. 184-188).

The auto-Gaussian model returns a spatial trend (autocorrelation) component which is calculated as follows (Cressie 2015, p. 564):

                            trend = rho * C * (Y - Mu).

This term can be extracted from a fitted auto-Gaussian model using the spatial method.

When applied to a fitted auto-Gaussian model, the residuals.geostan_fit method returns `de-trended' residuals, as follows:

                            residual = Y - Mu - trend.

The posterior predictive distribution (PPD) is returned in the parameter vector yrep. For the auto-Gaussian model, each sample from the PPD is a draw of n values from the multivariate Gaussian density function, with parameters Mu and Sigma. Draws from the PPD can be extracted from a fitted modeling using the posterior_predict method. The log likelihood of the data conditional on the model (log_lik, required for calculating waic) is calculated analogously, using the above parameterization of the multivariate normal density. Note, these quantities are only returned invert = TRUE, and may be costly to compute for large models.

family = poisson()

For family = poisson(), the model is specified as:

                            Y ~ Poisson(exp(offset + lambda))

                            lambda ~ MVGauss(Mu, Sigma)

                            Sigma = (I - rho C)^-1 * M * tau^2

Applying the fitted.geostan_fit method to this Poisson model will return exp(offset + lambda), fitted values on the scale of the outcome variable.

For Poisson models, the spatial method returns the parameter vector phi, which is:

                           phi = lambda - Mu.

This is equivalent to re-writing the model as:

                            Y ~ Poisson(exp(offset + Mu + phi))

                            phi ~ MVGauss(0, Sigma)

                            Sigma = (I - rho C)^-1 * M * tau^2.

Note that the behavior of the spatial method is slightly different for the CAR model than for the BYM model in stan_icar. The BYM model returns a spatial trend parameter that does not contain any uncorrelated/unstructured variation; phi, in the CAR model, contains a latent spatial trend and additional variation around it. If you would like to extract the latent/implicit spatial trend from phi, you can do so by calculating:

                            trend = rho * C * (phi - Mu).

This is not done automatically primarily because it is uncommon to do so.

The residuals.geostan_fit method returns:

                            residual = Y - exp(offset + lambda).

family = binomial()

For family = binomial(), the model is specified as:

                            Y ~ Binomial(N, theta)

                            logit(theta) ~ MVGauss(Mu, Sigma)

                            Sigma = (I - rho C)^-1 * M * tau^2

where outcome data Y are counts, N is the number of trials, and theta is the 'success' rate.

For fitted Binomial models, the spatial method will return the parameter vector phi, equivalent to:

                            phi = logit(theta) - Mu.

The fitted method applied to the Binomial model returns:

                           fitted = (theta * N)/N,

and the residuals.geostan_fit method returns:

                           residual = (Y/N) - fitted.

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(bayesplot)
library(sf)
options(mc.cores = parallel::detectCores())
data(sentencing)

C <- shape2mat(sentencing, style = "B")
car.dl <- prep_car_data(C, style = "WCAR")
log_e <- log(sentencing$expected_sents)
fit.car <- stan_car(sents ~ offset(log_e),
                    family = poisson(),
                    data = sentencing,
                    car_parts = car.dl,
                    refresh = 0
)

# posterior predictive distribution
yrep <- posterior_predict(fit.car, samples = 75)
y <- sentencing$sents
ppc_dens_overlay(y, yrep)

# examine posterior distributions of CAR parameters
plot(fit.car, pars = c("car_scale", "car_rho"))

# map the spatial autocorrelation term, phi
sp.trend <- spatial(fit.car)$mean
st_as_sf(sentencing) %>%
  ggplot() +
  geom_sf(aes(fill = sp.trend)) +
  scale_fill_gradient2()
 
# calculate log-standardized sentencing ratios (log-SSRs)
# (like Standardized Incidence Ratios: observed/exected case counts)
f <- fitted(fit.car)$mean
SSR <-  f / sentencing$expected_sents
log.SSR <- log( SSR, base = 2 )

st_as_sf(sentencing) %>%
 ggplot() +
 geom_sf(aes(fill = log.SSR)) +
 scale_fill_gradient2(
   midpoint = 0,
   name = NULL,
   breaks = seq(-3, 3, by = 0.5)
   ) +
 labs(title = "Log-Standardized Sentencing Ratios",
      subtitle = "log( Fitted/Expected ), base 2"
      ) +
 theme_void() +
 theme(
   legend.position = "bottom",
   legend.key.height = unit(0.35, "cm"),
   legend.key.width = unit(1.5, "cm")
 )
}