Spatial filtering

Fit a spatial regression model using eigenvector spatial filtering.

stan_esf(
  formula,
  slx,
  re,
  data,
  C,
  EV,
  ME = NULL,
  nsa = FALSE,
  threshold = 0.25,
  family = gaussian(),
  p0,
  prior = NULL,
  prior_intercept = NULL,
  prior_sigma = NULL,
  prior_rhs = NULL,
  prior_nu = NULL,
  prior_tau = NULL,
  centerx = FALSE,
  scalex = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 2000,
  refresh = 500,
  pars = NULL,
  control = list(adapt_delta = 0.99, max_treedepth = 15),
  silent = FALSE,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models are specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally). These will be pre-multiplied by a row-standardized spatial weights matrix and then added (prepended) to the design matrix. If and when setting priors for beta manually, remember to include priors for any SLX terms as well.

re
If the model includes a varying intercept specify the grouping variable here using formula syntax, as in ~ ID. The resulting random effects parameter returned is named alpha_re.

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

C
Spatial connectivity matrix which will be used to calculate eigenvectors, residual spatial autocorrelation as well as any user specified slx terms or spatial measurement error (ME) models; it will be row-standardized before calculating slx terms.

EV
A matrix of eigenvectors from any (transformed) connectivity matrix, presumably spatial. If provided, still also provide a spatial weights matrix C for other purposes. See make_EV and shape2mat.

ME
To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. Errors are assigned a Gaussian probability distribution and the modeled (true) covariate vector is assigned a Student's t model or, if ME$spatial = TRUE, an auto Gaussian (CAR) model. Elements of the list ME may include:

se

a dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounded

If any variables in se are bounded within some range (e.g. percentages ranging from zero to one hundred) provide a vector of zeros and ones indicating which columns are bounded. By default the lower bound will be 0 and the upper bound 100, for percentages.

bounds

A numeric vector of length two providing the upper and lower bounds, respectively, of any bounded variables.

spatial

Logical value indicating whether an auto Gaussian (i.e., conditional autoregressive (CAR)) model should be used for the covariates. Defaults to spatial = FALSE. If spatial = TRUE you must provide a connectivity matrix C. The specification of the auto Gaussian model is fixed such that all non-zero values in C will be converted to ones.

nsa
Include eigenvectors representing negative spatial autocorrelation? Default nsa = FALSE. Ignored if EV is provided.

threshold
Threshold for eigenvector MC value; eigenvectors with values below threshold will be excluded from the candidate set. Default threshold = 0.25; ignored if EV is provided.

family
The likelihood function for the outcome variable. Current options are family = gaussian(), student_t() and poisson(link = "log"), and binomial(link = "logit").

p0
Number of eigenvector coefficients expected to be far from zero. If this and prior_rhs are missing, Chun et al.'s (2016) formula will be used to fill this in; see exp_pars. The value of p0 is used to control the prior degree of sparsity in the model.

prior
A data.frame or matrix with location and scale parameters for Gaussian prior distributions on the model coefficients. Provide two columns---location and scale---and a row for each variable in their order of appearance in the model formula. Default priors are weakly informative relative to the scale of the data.

prior_intercept
A vector with location and scale parameters for a Gaussian prior distribution on the intercept; e.g. prior_intercept = c(0, 10).

prior_sigma
A vector with degrees of freedom, location and scale parameters for the half-Student's t prior on the residual standard deviation sigma. To use a half-Cauchy prior set degrees of freedom to one; e.g. prior_sigma = c(1, 0, 3).

prior_rhs
A named vector with the degrees of freedom and scale of the zero mean Student's t slab for regularizing large eigenvector coefficients, and the prior scale for the global shrinkage parameter; e.g. prior_rhs = c(slab_df = 15, slab_scale = 5, scale_global = 0.5). If p0 is provided and prior_rhs is missing, p0 will be used automatically to calculate scale_global, slab_df will be set to 15, and slab_scale will be set to be weakly informative automatically based on the scale of the data.

prior_nu
Set the parameters for the Gamma prior distribution on the degrees of freedom in the likelihood function when using family = student_t. Defaults to prior_nu = c(alpha = 3, beta = 0.2).

prior_tau
Set hyperparameters for the scale parameter of exchangeable random effects/varying intercepts (alpha_re). The random effects are given a normal prior with scale parameter alpha_tau. The latter is given a half-Student's t prior with default of 20 degrees of freedom, centered on zero and scaled to the data to be weakly informative. To adjust it use, e.g., prior_tau = c(df = 20, location = 0, scale = 5).

centerx
Should the covariates be centered prior to fitting the model? Defaults to FALSE.

scalex
Should the covariates be centered and scaled (divided by their standard deviation)? Defaults to FALSE.

prior_only
Draw samples from the prior distributions of parameters only.

chains
Number of MCMC chains to estimate. Default chains = 4.

iter
Number of samples per chain. Default iter = 2000.

refresh
Stan will print the progress of the sampler every refresh number of samples. Defaults to 500; set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model. Parameters from the RHS priors include tau (the global shrinkage parameter) and lambda (the local shrinkage parameter).

control
A named list of parameters to control the sampler's behavior. See stan for details. The defaults are the same rstan::stan except that adapt_delta is raised to .99 and max_treedepth = 15.

silent
If TRUE, suppress printed messages including prior specifications and Stan sampling progress (i.e. refresh=0). Stan's error and warning messages will still print.

...
Other arguments passed to sampling.

Source

Chun, Y., D. A. Griffith, M. Lee and P. Sinha (2016). Eigenvector selection with stepwise regression techniques to construct eigenvector spatial filters. Journal of Geographical Systems, 18(1), 67-85. 10.1007/s10109-015-0225-3

Dray, S., P. Legendre & P. R. Peres-Neto (2006). Spatial modelling: a comprehensive framework for principal coordinate analysis of neighbour matrices (PCNM). Ecological Modeling, 196(3-4), 483-493.

Donegan, C., Y. Chun and A. E. Hughes (2020). Bayesian Estimation of Spatial Filters with Moran’s Eigenvectors and Hierarchical Shrinkage Priors. Spatial Statistics 38. <doi.org/10.1016/j.spasta.2020.100450>

Griffith, Daniel A., and P. R. Peres-Neto (2006). Spatial modeling in ecology: the flexibility of eigenfunction spatial analyses. Ecology 87(10), 2603-2613.

Griffith, D., and Y. Chun (2014). Spatial autocorrelation and spatial filtering, Handbook of Regional Science. Fischer, MM and Nijkamp, P. eds.

Griffith, D., Chun, Y. and Li, B. (2019). Spatial Regression Analysis Using Eigenvector Spatial Filtering. Elsevier.

Piironen, J and A. Vehtari (2017). Sparsity information and regularization in the horseshoe and other shrinkage priors. In Electronic Journal of Statistics, 11(2):5018-5051. https://projecteuclid.org/euclid.ejs/1513306866

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame

diagnostic

Widely Applicable Information Criteria (WAIC) with crude measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and residual spatial autocorrelation (Moran coefficient of the residuals). Residuals are relative to the mean posterior fitted values.

data

a data frame containing the model data

EV

A matrix of eigenvectors created with w and geostan::make_EV

C

The spatial weights matrix used to construct EV

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including ESF component)

slx

The slx formula

re

A list containing re, the random effects (varying intercepts) formula if provided, and data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

scale_params

A list with the center and scale parameters returned from the call to base::scale on the model matrix. If centerx = FALSE and scalex = FALSE then it is an empty list.

ME

The ME data list, if one was provided by the user for measurement error models.

spatial

A data frame with the name of the spatial component parameter ("esf") and method ("ESF")

stanfit

an object of class stanfit returned by rstan::stan

Details

Eigenvector spatial filtering is extensivly covered in Griffith et al. (2019). This function implements the methodology introduced in Donegan et al. (2020), drawing on the Piironen and Vehtari's (2017) regularized horseshoe prior. ESF models take the spectral decomposition of the doubly-centered spatial connectivity matrix C, which appears in the numerator of the Moran coefficient (MC),a measure of spatial autocorrelation. The resulting eigenvectors EV are mutually orthogonal and uncorrelated map patterns. ESF methodology is premised on our ability to decompose spatial autocorrelation into a linear combination of various patterns, typically at different scales (such as local, regional, and global trends). SA in the outcome variable is then absorbed by the spatial filter EV * beta_ev, where beta_ev is a vector of coefficients.

ESF decomposes the data into a global mean alpha, global patterns contributed by covariates X * beta, spatial trends EV * beta_ev, and residual variation. Thus

Mu = intercept + X * beta + EV * beta_ev

and for the Gaussian model,

Y \~ Gauss(Mu, sigma)

The model can easily be extended to the space-time domain; see shape2mat to specify a space-time connectivity matrix.

The coefficients beta_ev are assigned the regularized horseshoe prior (Piironen and Vehtari, 2017), resulting in a relatively sparse model specification. In addition, numerous eigenvectors are automatically dropped because they represent trace amounts of spatial autocorrelation (this is controlled by the threshold argument). By default, stan_esf will drop all eigenvectors representing negative spatial autocorrelation patterns. You can change this behavior using the nsa argument.

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(rstan)
library(bayesplot)
library(ggplot2)
library(sf)
options(mc.cores = parallel::detectCores()) 

data(ohio)
C <- shape2mat(ohio, "B")

fit <- stan_esf(gop_growth ~ historic_gop + college_educated,
                data = ohio,
                C = C)

# use rstan for sampler diagnostics
stan_rhat(fit$stanfit)
# see spatial autocorrelation diagnostics
sp_diag(fit, ohio)

## summary of estimates
plot(fit, pars = c("beta", "intercept", "sigma"))

## map the spatial filter
ohio$esf <- spatial(fit)$mean
ggplot(ohio) +
      geom_sf(aes(fill = esf)) +
      scale_fill_gradient2() +
      theme_void()

## make use of some additional features
 fit <- stan_esf(gop_growth ~ historic_gop + college_educated,
                slx = ~ historic_gop + college_educated,
                ME = list(se = data.frame(college_educated = ohio$college_educated.se),
                          spatial = TRUE),
                data = ohio,
                centerx = TRUE,
                C = C,
                family = student_t())

}