Spatial filtering

Fit a spatial regression model using eigenvector spatial filtering (ESF).

stan_esf(
  formula,
  slx,
  re,
  data,
  C,
  EV,
  ME = NULL,
  nsa = FALSE,
  threshold = 0.25,
  family = gaussian(),
  p0,
  prior = NULL,
  centerx = FALSE,
  scalex = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 2000,
  refresh = 500,
  pars = NULL,
  control = list(adapt_delta = 0.99, max_treedepth = 15),
  silent = FALSE,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models are specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, slx = ~ x1 + x2 (the implicit intercept term in this formula is removed internally). These will be pre-multiplied by a row-standardized spatial weights matrix and then added (prepended) to the design matrix. If and when setting priors for beta manually, remember to include priors for any SLX terms as well.

re
If the model includes a varying intercept specify the grouping variable here using formula syntax, as in ~ ID. The resulting random effects parameter returned is named alpha_re. The scale parameter for this term is named alpha_tau. That is, alpha_re ~ N(0, alpha_tau), alpha_tau ~ Student_t(d.f., location, scale).

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

C
Spatial connectivity matrix which will be used to calculate eigenvectors, residual spatial autocorrelation as well as any user specified slx terms or spatial measurement error (ME) models; it will be row-standardized before calculating slx terms.

EV
A matrix of eigenvectors from any (transformed) connectivity matrix, presumably spatial. If provided, still also provide a spatial weights matrix C for other purposes. See make_EV and shape2mat.

ME
To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. Errors are assigned a Gaussian probability distribution and the modeled (true) covariate vector is assigned a Student's t model or, if ME$spatial = TRUE, an auto Gaussian (CAR) model. Elements of the list ME may include:

se

a dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounded

If any variables in se are bounded within some range (e.g. percentages ranging from zero to one hundred) provide a vector of zeros and ones indicating which columns are bounded. By default the lower bound will be 0 and the upper bound 100, for percentages.

bounds

A numeric vector of length two providing the upper and lower bounds, respectively, of any bounded variables.

spatial

Logical value indicating whether an auto Gaussian (conditional autoregressive (CAR)) model should be used for the covariates (see stan_car for CAR model details). For stan_esf, this defaults to spatial = TRUE. If spatial = TRUE you must provide car_parts (see below).

car_parts

A list of data for the CAR model, as returned by prep_car_data() (be sure to return the matrix C, by using the argument cmat = TRUE). Only required if spatial=TRUE.

nsa
Include eigenvectors representing negative spatial autocorrelation? Default nsa = FALSE. Ignored if EV is provided.

threshold
Threshold for eigenvector MC value; eigenvectors with values below threshold will be excluded from the candidate set. Default threshold = 0.25; ignored if EV is provided.

family
The likelihood function for the outcome variable. Current options are family = gaussian(), student_t() and poisson(link = "log"), and binomial(link = "logit").

p0
Number of eigenvector coefficients expected to be far from zero. If this and prior_rhs are missing, Chun et al.'s (2016) formula will be used to fill this in; see exp_pars. The value of p0 is used to control the prior degree of sparsity in the model.

prior
A named list of parameters for prior distributions. User-defined priors can be assigned to the following parameters:

intercept

The intercept is assigned a Gaussian prior distribution; provide a numeric vector with location and scale parameters; e.g. prior = list(intercept = c(location = 0, scale = 10)) to assign a Gaussian prior with mean zero and variance 10^2.

beta

Regression coefficients are assigned Gaussian prior distributions. Provide a data.frame with two columns (location and scale).The order of variables must follow their order of appearance in the model formula; e.g., if the formula is y ~ x1 + x2, then providing `prior = list(beta = data.frame(location = c(0, 0), scale = c(1, 1))) will assign standard normal prior distributions to both coefficients.Note that if you also use slx terms (spatially lagged covariates), then you have to provide priors for them; slx terms precede other terms in the model formula. For example, providing of y ~ x1 + x2 with slx = ~ x1 + x2 is the same as providing formula = y ~ Wx1 + Wx2 + x1 + x2, where Wx1 indicates the spatially lagged covariate.

sigma

The scale parameter for Gaussian and Student's t likelihood models is assigned a half-Student's t prior distribution. Thus, provide values for the degrees of freedom, location, and scale parameters; e.g., prior = list(sigma = c(df = 10, location = 0, scale = 5)) for a prior centered on zero with scale of 5 and 10 degrees of freedom. The half-Student's t prior for sigma is constrained to be positive.

nu

The degrees of freedom for the Student's t likelihood model is assigned a gamma prior distribution. The default prior is prior = list(nu = c(alpha = 3, beta = 0.2)).

tau

The scale parameter for random effects, or varying intercepts, terms. This scale parameter, tau, is assigned a half-Student's t prior. To set this, use, e.g., prior = list(tau = c(df = 20, location = 0, scale = 20)).

rhs

This is for controlling the regularized horseshoe (RHS) prior, which is assigned to the eigenvector coefficients. Provide a named vector of length three. The RHS prior has two parts to consider. First is the 'slab' that is used to regularize large coefficient estimates; this is a zero-mean Student's t model, the user provides the degrees of freedom and scale parameters. The second component is the global shrinkage parameter, controlling the degree of shrinkage; provide a value nearer to zero if you wish to make the model more sparse. To allow the spatial filter to account for a greater amount of spatial autocorrelation (i.e., if you find the residuals contain spatial autocorrelation), increase the global scale parameter up to one. E.g., prior = list(rhs = c(slab_df = 15, slab_scale = 5, scale_global = 0.5)).

centerx
To center predictors, provide either a logical value (TRUE, FALSE) or numeric-alike vector of length equal to the number of columns of ‘x’, where ‘numeric-alike’ means that ‘as.numeric(.)’ will be applied successfully if ‘is.numeric(.)’ is not true. Passed to scale.

scalex
To scale predictors, provide either a logical value or a numeric-alike vector of length equal to the number of columns of ‘x’. Passed to scale.

prior_only
Draw samples from the prior distributions of parameters only.

chains
Number of MCMC chains to estimate. Default chains = 4.

iter
Number of samples per chain. Default iter = 2000.

refresh
Stan will print the progress of the sampler every refresh number of samples. Defaults to 500; set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model. Parameters from the RHS priors include tau (the global shrinkage parameter) and lambda (the local shrinkage parameter).

control
A named list of parameters to control the sampler's behavior. See stan for details. The defaults are the same rstan::stan except that adapt_delta is raised to .99 and max_treedepth = 15.

silent
If TRUE, suppress printed messages including prior specifications and Stan sampling progress (i.e. refresh=0). Stan's error and warning messages will still print.

...
Other arguments passed to sampling.

Source

Chun, Y., D. A. Griffith, M. Lee and P. Sinha (2016). Eigenvector selection with stepwise regression techniques to construct eigenvector spatial filters. Journal of Geographical Systems, 18(1), 67-85. 10.1007/s10109-015-0225-3

Dray, S., P. Legendre & P. R. Peres-Neto (2006). Spatial modelling: a comprehensive framework for principal coordinate analysis of neighbour matrices (PCNM). Ecological Modeling, 196(3-4), 483-493.

Donegan, C., Y. Chun and A. E. Hughes (2020). Bayesian Estimation of Spatial Filters with Moran’s Eigenvectors and Hierarchical Shrinkage Priors. Spatial Statistics 38. doi.org/10.1016/j.spasta.2020.100450

Griffith, Daniel A., and P. R. Peres-Neto (2006). Spatial modeling in ecology: the flexibility of eigenfunction spatial analyses. Ecology 87(10), 2603-2613.

Griffith, D., and Y. Chun (2014). Spatial autocorrelation and spatial filtering, Handbook of Regional Science. Fischer, MM and Nijkamp, P. eds.

Griffith, D., Chun, Y. and Li, B. (2019). Spatial Regression Analysis Using Eigenvector Spatial Filtering. Elsevier.

Piironen, J and A. Vehtari (2017). Sparsity information and regularization in the horseshoe and other shrinkage priors. In Electronic Journal of Statistics, 11(2):5018-5051. https://projecteuclid.org/euclid.ejs/1513306866

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame

diagnostic

Widely Applicable Information Criteria (WAIC) with crude measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and residual spatial autocorrelation (Moran coefficient of the residuals). Residuals are relative to the mean posterior fitted values.

data

a data frame containing the model data

EV

A matrix of eigenvectors created with w and geostan::make_EV

C

The spatial weights matrix used to construct EV

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including ESF component)

slx

The slx formula

re

A list containing re, the random effects (varying intercepts) formula if provided, and data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

scale_params

A list with the center and scale parameters returned from the call to base::scale on the model matrix. If centerx = FALSE and scalex = FALSE then it is an empty list.

ME

The ME data list, if one was provided by the user for measurement error models.

spatial

A data frame with the name of the spatial component parameter ("esf") and method ("ESF")

stanfit

an object of class stanfit returned by rstan::stan

Details

Eigenvector spatial filtering (ESF) is extensivly covered in Griffith et al. (2019). This function implements the methodology introduced in Donegan et al. (2020), drawing on the Piironen and Vehtari's (2017) regularized horseshoe prior.

ESF models take the spectral decomposition of a transformed spatial connectivity matrix, C. The resulting eigenvectors, EV, are mutually orthogonal and uncorrelated map patterns. ESF decomposes spatial autocorrelation into a linear combination of various patterns, typically at different scales (such as local, regional, and global trends). By adding a spatial filter to a regression model, any spatial autocorrelation is shifted from the residuals to the spatial filter. The spatail filter is EV * beta_ev, where beta_ev is a vector of coefficients.

ESF decomposes the data into a global mean, alpha, global patterns contributed by covariates, X * beta, spatial trends, EV * beta_ev, and residual variation. Thus, for family=gaussian(),

                         Y ~ Gauss(alpha + X * beta + EV * beta_ev, sigma).

An ESF component can be incorporated into the linear predictor of any generalized linear model. For example, a spatial Poisson model for rare disease incidence may be specified as follows:

                          Y ~ Poisson(exp(offset + Mu))
                          Mu = alpha + A + EV * beta_ev
                          A ~ Guass(0, tau)
                          tau ~ student(20, 0, 2)
                          beta_ev ~ horseshoe(.)

The spatial method will return EV * beta.

The model can also be extended to the space-time domain; see shape2mat to specify a space-time connectivity matrix.

The coefficients beta_ev are assigned the regularized horseshoe prior (Piironen and Vehtari, 2017), resulting in a relatively sparse model specification. In addition, numerous eigenvectors are automatically dropped because they represent trace amounts of spatial autocorrelation (this is controlled by the threshold argument). By default, stan_esf will drop all eigenvectors representing negative spatial autocorrelation patterns. You can change this behavior using the nsa argument.

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(sf)
library(bayesplot)
data(sentencing)

# spatial weights matrix with binary coding scheme
C <- shape2mat(sentencing, style = "B")

# log-expected number of sentences
## expected counts are based on county racial composition and mean sentencing rates
log_e <- log(sentencing$expected_sents)

# fit spatial Poisson model with ESF + unstructured 'random effects'
fit.esf <- stan_esf(sents ~ offset(log_e),
                   re = ~ name,
                   family = poisson(),
                   data = sentencing,
                   C = C,
                   refresh = 0
)

# spatial diagnostics 
sp_diag(fit.esf, sentencing)

# plot marginal posterior distributions of beta_ev (eigenvector coefficients)
plot(fit.esf, pars = "beta_ev")

# plot the marginal posterior distributions of the spatial filter (ESF * beta_ev)
plot(fit.esf, pars = "beta_ev")

# posterior predictive distribution
yrep <- posterior_predict(fit.esf, samples = 75)
y <- sentencing$sents
bayesplot::ppc_dens_overlay(y, yrep) 

# map the spatial filter
sp.filter <- spatial(fit.esf)$mean
st_as_sf(sentencing) %>%
 ggplot() +
 geom_sf(aes(fill = sp.filter)) +
 scale_fill_gradient2()

# calculate log-standardized sentencing ratios (log-SSRs)
f <- fitted(fit.esf)$mean
SSR <-  f / sentencing$expected_sents
log.SSR <- log( SSR, base = 2 )

# map the log-SSRs
st_as_sf(sentencing) %>%
 ggplot() +
 geom_sf(aes(fill = log.SSR)) +
 scale_fill_gradient2(
   midpoint = 0,
   name = NULL,
   breaks = seq(-3, 3, by = 0.5)
 ) +
 labs(title = "Log-Standardized Sentencing Ratios",
      subtitle = "log( Fitted/Expected ), base 2"
 ) +
 theme_void() +
 theme(
   legend.position = "bottom",
   legend.key.height = unit(0.35, "cm"),
   legend.key.width = unit(1.5, "cm")
 )

}