Generalized linear models

Fit a generalized linear model.

stan_glm(
  formula,
  slx,
  re,
  data,
  C,
  family = gaussian(),
  prior = NULL,
  ME = NULL,
  centerx = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 2000,
  refresh = 1000,
  pars = NULL,
  control = NULL,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models are specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally). When setting priors for beta, remember to include priors for any SLX terms.

re
To include a varying intercept (or "random effects") term, alpha_re, specify the grouping variable here using formula syntax, as in ~ ID. Then, alpha_re is a vector of parameters added to the linear predictor of the model, and:

       alpha_re ~ N(0, alpha_tau)
       alpha_tau ~ Student_t(d.f., location, scale).

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

C
Optional spatial connectivity matrix which will be used to calculate residual spatial autocorrelation as well as any user specified slx terms; it will automatically be row-standardized before calculating slx terms. See shape2mat.

family
The likelihood function for the outcome variable. Current options are poisson(link = "log"), binomial(link = "logit"), student_t(), and the default gaussian().

prior
A named list of parameters for prior distributions (see priors):

intercept

The intercept is assigned a Gaussian prior distribution (see normal

.
beta

Regression coefficients are assigned Gaussian prior distributions. Variables must follow their order of appearance in the model formula. Note that if you also use slx terms (spatially lagged covariates), and you use custom priors for beta, then you have to provide priors for the slx terms. Since slx terms are prepended to the design matrix, the prior for the slx term will be listed first.

sigma

For family = gaussian() and family = student_t() models, the scale parameter, sigma, is assigned a (half-) Student's t prior distribution. The half-Student's t prior for sigma is constrained to be positive.

nu

nu is the degrees of freedom parameter in the Student's t likelihood (only used when family = student_t()). nu is assigned a gamma prior distribution. The default prior is prior = list(nu = gamma(alpha = 3, beta = 0.2)).

tau

The scale parameter for random effects, or varying intercepts, terms. This scale parameter, tau, is assigned a half-Student's t prior. To set this, use, e.g., prior = list(tau = student_t(df = 20, location = 0, scale = 20)).

ME
To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list of priors. See the Details section below for more information. Elements of the ME list may include:

se

A required dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounds

An optional numeric vector of length two providing the upper and lower bounds, respectively, of the variables. If not provided, they will be set to c(-Inf, Inf) (i.e., unbounded). Common usages include keeping percentages between zero and one hundred or proportions between zero and one.

car_parts

A list of data for the CAR model, as returned by prep_car_data. If not provided, a non-spatial Student's t model will be used instead of the CAR model.

prior

Optionally provide a named list of prior distributions for the measurement error model(s) (see priors). If none are provided, default priors will be assigned and printed to the console.

df

If using a non-spatial ME model, the degrees of freedom for the Student's t model is assigned a gamma prior with default parameters of gamma(alpha = 3, beta = 0.2). Provide values for each covariate in se, listing the values in the same order as the columns of se.

car_rho

The CAR model also has a spatial autocorrelation parameter, rho, which is assigned a uniform prior distribution. You can set the boundaries of the prior with: ME$prior$car_rho <- uniform(lower_bound, upper_bound). You must specify values that are within the permissible range of values for rho. This range is automatically printed to the console by prep_car_data.

location

The prior for the location parameter, mu, is Gaussian (the default being normal(location = 0, scale = 100)). To adjust the prior distributions, provide values for each covariate in se, listing the values in the same order as the columns of se.

scale

The prior for the scale parameters is Student's t, and the default is student_t(df = 10, location = 0, scale = 40). To adjust, provide values for each covariate in se, listing the values in the same order as the columns of se.

centerx
To center predictors on their mean values, use centerx = TRUE. If centerx is a numeric vector, predictors will be centered on the values provided. This argument is passed to scale.

prior_only
Draw samples from the prior distributions of parameters only.

chains
Number of MCMC chains to estimate.

iter
Number of samples per chain.

refresh
Stan will print the progress of the sampler every refresh number of samples; set refresh=0 to silence this.

pars
Specify any additional parameters you'd like stored from the Stan model.

control
A named list of parameters to control the sampler's behavior. See stan for details.

...
Other arguments passed to sampling. For multi-core processing, you can use cores = parallel::detectCores(), or run options(mc.cores = parallel::detectCores()) first.

Source

Donegan, Connor and Chun, Yongwan and Griffith, Daniel A. (2021). Modeling community health with areal data: Bayesian inference with survey standard errors and spatial structure. Int. J. Env. Res. and Public Health 18 (13): 6856. DOI: 10.3390/ijerph18136856 Data and code: https://github.com/ConnorDonegan/survey-HBM.

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame

diagnostic

Widely Applicable Information Criteria (WAIC) with a measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and mean residual spatial autocorrelation as measured by the Moran coefficient.

stanfit

an object of class stanfit returned by rstan::stan

data

a data frame containing the model data

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including ESF component)

slx

The slx formula

re

A list containing re, the random effects (varying intercepts) formula if provided, and Data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

x_center

If covariates are centered internally (i.e., centerx is not FALSE), then x_centers is the numeric vector of values on which the covariates were centered.

ME

The ME data list, if one was provided by the user for measurement error models.

spatial

NA, slot is maintained for use in geostan_fit methods.

Details

Fit a generalized linear model using the R formula interface. Default prior distributions are designed to be weakly informative relative to the data. Much of the functionality intended for spatial models, such as the ability to add spatially lagged covariates and observational error models, are also available in stan_glm. All of geostan's spatial models build on top of the same Stan code used in stan_glm.

Poisson models and disease mapping

In spatial statistics, Poisson models are often used to calculate incidence rates (mortality rates, or disease incidence rates) for administrative areas like counties or census tracts. If Y are counts of cases, and P are populations at risk, then the crude rates are Y/P. For a Poisson model, Y is the outcome and the user should provide an offset value of log(P). For such a case, disease incidence across the collection of areas could be modeled as:

        Y ~ Poisson(exp(offset + Mu))
        Mu = alpha + A
        A ~ Guass(0, tau)
        tau ~ student(20, 0, 2),

where alpha is the mean log-risk (incidence rate) and A is a vector of (so-called) random effects, which enable partial pooling of information across observations. See the example section of this document for a demonstration (where the denominator of the outcome is the expected count, rather than population at risk).

Spatially lagged covariates (SLX)

The slx argument is a convenience function for including SLX terms. For example,

stan_glm(y ~ x1 + x2, slx = ~ x1, ...)

is a shortcut for

stan_glm(y ~ I(W %*% x1) + x1 + x2, ...)

where W is a row-standardized spatial weights matrix (see shape2mat). SLX terms will always be prepended to the design matrix, as above, which is important to know when setting prior distributions for regression coefficients.

For measurement error (ME) models, the SLX argument is the only way to include spatially lagged covariates since the SLX term needs to be re-calculated on each iteration of the MCMC algorithm.

Measurement error (ME) models

The ME models are designed for surveys with spatial sampling designs, such as the American Community Survey (ACS) estimates. With estimates, x, and their standard errors, se, the ME models have one of the the following two specifications, depending on the user input:

       x ~ Gauss(x_true, se)
       x_true ~ MVGauss(mu, Sigma)
       Sigma = (I - rho * C)^(-1) M * tau^2
       mu ~ Gauss(0, 100)
       tau ~ student_t(10, 0, 40)
       rho ~ uniform(lower_bound, upper_bound)

where the covariance matrix, Sigma, has the conditional autoregressive specification, and tau is the scale parameter. If ME$car_parts is not provided by the user, then a non-spatial model will be used instead:

       x ~ Gauss(x_true, se)
       x_true ~ student_t(df, mu, sigma)
       df ~ gamma(3, 0.2)
       mu ~ Gauss(0, 100)
       sigma ~ student_t(10, 0, 40)

The observational error is the difference between the survey estimate and x_true, the actual value of the variable during the period of the survey.

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(sf)
data(sentencing)

sentencing$log_e <- log(sentencing$expected_sents)
fit.pois <- stan_glm(sents ~ offset(log_e),
                     re = ~ name,
                     family = poisson(),
                     data = sentencing
 )

# MCMC diagnostics plot: Rhat values should all by very near 1
rstan::stan_rhat(fit.pois$stanfit)
 # see effective sample size for all parameters and generated quantities
 # (including residuals, predicted values, etc.)
rstan::stan_ess(fit.pois$stanfit)
# or for a particular parameter
rstan::stan_ess(fit.pois$stanfit, "alpha_re")

# Spatial autocorrelation/residual diagnostics
sp_diag(fit.pois, sentencing)

## Posterior predictive distribution                                       
library(bayesplot)
yrep <- posterior_predict(fit.pois, samples = 75)
y <- sentencing$sents
ppc_dens_overlay(y, yrep)
}