Generalized linear models

Fit a generalized linear model with (optional) exchangeable random effects.

stan_glm(
  formula,
  slx,
  re,
  data,
  ME = NULL,
  C,
  EV,
  family = gaussian(),
  prior = NULL,
  prior_intercept = NULL,
  prior_sigma = NULL,
  prior_nu = NULL,
  prior_tau = NULL,
  centerx = FALSE,
  scalex = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 4000,
  refresh = 1000,
  pars = NULL,
  control = list(adapt_delta = 0.95, max_treedepth = 15),
  silent = FALSE,
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models are specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally); you must also provide C when including slx. Specified covariates will be pre-multiplied by a row-standardized spatial weights matrix and then added (prepended) to the design matrix. If and when setting priors for beta manually, remember to include priors for any SLX terms as well.

re
If the model includes a varying intercept term (or "spatially unstructured random effect") specify the grouping variable here using formula syntax, as in ~ ID. The resulting random effects parameter returned is named alpha_re.

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

ME
To model observational error (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. Errors are assigned a Gaussian probability distribution and the modeled (true) covariate vector is assigned a Student's t model with optional spatially varying mean. Elements of the list ME may include:

se

a dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounded

If any variables in se are bounded within some range (e.g. percentages ranging from zero to one hundred) provide a vector of zeros and ones indicating which columns are bounded. By default the lower bound will be 0 and the upper bound 100, for percentages.

bounds

A numeric vector of length two providing the upper and lower bounds, respectively, of the bounded variables. Defaults to bounds = c(0, 100).

spatial

Logical value indicating if the models for covariates should include a spatially varying mean (using an eigenvector spatial filter). Defaults to spatial = FALSE. If spatial = TRUE and you do not provide both ME$prior_rhs and EV then you must provide a connectivity matrix C.

prior_rhs

Optional prior parameters for the regularized horseshoe (RHS) prior used for the ESF data model; only used if ME$spatial = TRUE. The RHS prior is used for the eigenvector spatial filter (ESF), as in stan_esf. Must be a named list containing vectors slab_df, slab_scale, scale_global, and varname. The character vector varname indicates the order of the other parameters (by name).

C
Optional spatial connectivity matrix which will be used to calculate residual spatial autocorrelation as well as any user specified slx or spatial ME terms; it will be row-standardized before calculating slx terms.

EV
A matrix of eigenvectors from any (transformed) connectivity matrix (presumably spatial). See make_EV and shape2mat.

family
The likelihood function for the outcome variable. Current options are poisson(link = "log"), binomial(link = "logit"), student_t(), and the default gaussian().

prior
A data.frame or matrix with location and scale parameters for Gaussian prior distributions on the model coefficients. Provide two columns---location and scale---and a row for each variable in their order of appearance in the model formula. Default priors are weakly informative relative to the scale of the data.

prior_intercept
A vector with location and scale parameters for a Gaussian prior distribution on the intercept; e.g. prior_intercept = c(0, 10).

prior_sigma
A vector with degrees of freedom, location and scale parameters for the half-Student's t prior on the residual standard deviation sigma. To use a half-Cauchy prior set degrees of freedom to one; e.g. prior_sigma = c(1, 0, 3).

prior_nu
Set the parameters for the Gamma prior distribution on the degrees of freedom in the likelihood function when using family = student_t. Defaults to prior_nu = c(alpha = 3, beta = 0.2).

prior_tau
Set hyperparameters for the scale parameter of exchangeable random effects/varying intercepts. The random effects are given a normal prior with scale parameter alpha_tau. The latter is given a half-Student's t prior with default of 20 degrees of freedom, centered on zero and scaled to the data to be weakly informative. To adjust it use, e.g., prior_tau = c(df = 20, location = 0, scale = 20).

centerx
Should the covariates be centered prior to fitting the model? Defaults to FALSE.

scalex
Should the covariates be centered and scaled (divided by their standard deviation)? Defaults to FALSE.

prior_only
Draw samples from the prior distributions of parameters only.

chains
Number of MCMC chains to estimate.

iter
Number of samples per chain.

refresh
Stan will print the progress of the sampler every refresh number of samples; set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model.

control
A named list of parameters to control the sampler's behavior. See stan for details. The defaults are the same rstan::stan except that adapt_delta is raised to 0.95 and max_treedepth = 15.

silent
If TRUE, suppress printed messages including prior specifications and Stan sampling progress (i.e. refresh=0). Stan's error and warning messages will still print.

...
Other arguments passed to sampling. For multi-core processing, you can use cores = parallel::detectCores(), or run options(mc.cores = parallel::detectCores()) first.

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame

diagnostic

Widely Applicable Information Criteria (WAIC) with crude measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and residual spatial autocorrelation (Moran coefficient of the residuals). Residuals are relative to the mean posterior fitted values.

stanfit

an object of class stanfit returned by rstan::stan

data

a data frame containing the model data

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including ESF component)

slx

The slx formula

re

A list containing re, the random effects (varying intercepts) formula if provided, and Data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

scale_params

A list with the center and scale parameters returned from the call to base::scale on the model matrix. If centerx = FALSE and scalex = FALSE then it is an empty list.

ME

The ME data list, if one was provided by the user for measurement error models.

spatial

NA, slot is maintained for use in geostan_fit methods.

Details

Fit a generalized linear model using the R formula interface. Default prior distributions are designed to be weakly informative relative to the data. Much of the functionality intended for spatial models, such as the ability to add spatially lagged covariates and observational error models, are also available in stan_glm. All of geostan's spatial models build on top of the same Stan code used in stan_glm.

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(sf)
data(sentencing)

sentencing$log_e <- log(sentencing$expected_sents)
fit.pois <- stan_glm(sents ~ offset(log_e),
                     re = ~ name,
                     family = poisson(),
                     data = sentencing,
                    silent = TRUE
 )

# diagnostics plot: Rhat values should all by very near 1
library(rstan)
rstan::stan_rhat(fit.pois$stanfit)
 # see effective sample size for all parameters and generated quantities
 # (including residuals, predicted values, etc.)
rstan::stan_ess(fit.pois$stanfit)
# or for a particular parameter
rstan::stan_ess(fit.pois$stanfit, "alpha_re")

# spatial autocorrelation/residual diagnostics
sp_diag(fit.pois, sentencing)

library(bayesplot)
yrep <- posterior_predict(fit.pois, samples = 75)
y <- sentencing$sents
ppc_dens_overlay(y, yrep)
}