Generalized linear models

Fit a generalized linear model.

stan_glm(
  formula,
  slx,
  re,
  data,
  C,
  ME = NULL,
  family = gaussian(),
  prior = NULL,
  centerx = FALSE,
  prior_only = FALSE,
  chains = 4,
  iter = 2000,
  refresh = 1000,
  pars = NULL,
  control = list(adapt_delta = 0.95, max_treedepth = 12),
  ...
)

Arguments

formula
A model formula, following the R formula syntax. Binomial models are specified by setting the left hand side of the equation to a data frame of successes and failures, as in cbind(successes, failures) ~ x.

slx
Formula to specify any spatially-lagged covariates. As in, ~ x1 + x2 (the intercept term will be removed internally); you must also provide C when including slx. Specified covariates will be pre-multiplied by a row-standardized spatial weights matrix and then added (prepended) to the design matrix. If and when setting priors for beta manually, remember to include priors for any SLX terms as well.

re
If the model includes a varying intercept term (or "spatially unstructured random effect") specify the grouping variable here using formula syntax, as in ~ ID. The resulting random effects parameter returned is named alpha_re, and the associated scale parameter is alpha_tua. That is, alpha_re ~ N(0, alpha_tau), alpha_tau ~ Student_t(d.f., location, scale).

data
A data.frame or an object coercible to a data frame by as.data.frame containing the model data.

C
Optional spatial connectivity matrix which will be used to calculate residual spatial autocorrelation as well as any user specified slx or spatial measurement error (ME) terms; it will automatically be row-standardized before calculating slx terms.

ME
To model observational uncertainty (i.e. measurement or sampling error) in any or all of the covariates, provide a named list. Errors are assigned a Gaussian probability distribution and the modeled (true) covariate vector is assigned a Student's t model or, if ME$spatial = TRUE, an auto Gaussian (CAR) model. Elements of the list ME may include:

se

a dataframe with standard errors for each observation; columns will be matched to the variables by column names. The names should match those from the output of model.matrix(formula, data).

bounded

If any variables in se are bounded within some range (e.g. percentages ranging from zero to one hundred) provide a vector of zeros and ones indicating which columns are bounded. By default the lower bound will be 0 and the upper bound 100, for percentages.

bounds

A numeric vector of length two providing the upper and lower bounds, respectively, of any bounded variables.

prior

Provide parameter values for the prior distributions of the measurement error model(s). The spatial conditional autoregressive (CAR) and non-spatial student's t models both contain a location parameter (the mean) and a scale parameter, each of which require prior distributions. If none are provided, default priors will be assigned and printed to the console.To set these priors to custom values, provide a named list with items location and scale (you must provide both). The prior for the location parameter is Gaussian, the default being Gauss(0, 100). To change these values, provide a data.frame with columns named location and scale. Provide prior specifications for each covariate in se; list the priors in the same order as the columns of se.The prior for the scale parameters is Student's t, and the default is Student_t(10, 0, 40). The degrees of freedom (10) and mean (zero) are fixed, but you can alter the scale by providing a vector of values in the same order as the columns of se.For example, ME$prior <- list(location = data.frame(location = c(0, 0), scale = c(100, 100)), scale = c(40, 40)).

spatial

Logical value indicating whether an auto Gaussian (i.e., conditional autoregressive (CAR)) model should be used for the covariates. For stan_glm, this defaults to spatial = FALSE. If spatial = TRUE you must provide car_parts (see below).

car_parts

A list of data for the CAR model, as returned by prep_car_data (be sure to return the matrix C, by using the argument cmat = TRUE). Only required if spatial=TRUE.

family
The likelihood function for the outcome variable. Current options are poisson(link = "log"), binomial(link = "logit"), student_t(), and the default gaussian().

prior
A named list of parameters for prior distributions. User-defined priors can be assigned to the following parameters:

intercept

The intercept is assigned a Gaussian prior distribution; provide a numeric vector with location and scale parameters; e.g. prior = list(intercept = c(location = 0, scale = 10)) to assign a Gaussian prior with mean zero and variance 10^2.

beta

Regression coefficients are assigned Gaussian prior distributions. Provide a data.frame with two columns (location and scale).The order of variables must follow their order of appearance in the model formula; e.g., if the formula is y ~ x1 + x2, then providing `prior = list(beta = data.frame(location = c(0, 0), scale = c(1, 1))) will assign standard normal prior distributions to both coefficients.Note that if you also use slx terms (spatially lagged covariates), then you have to provide priors for them; slx terms precede other terms in the model formula. For example, providing of y ~ x1 + x2 with slx = ~ x1 + x2 is the same as providing formula = y ~ I(W \%*\% x1) + I(W \%*\% x2) + x1 + x2, where W is a row-standardized spatial weights matrix (see shape2mat).

sigma

The scale parameter for Gaussian and Student's t likelihood models is assigned a half-Student's t prior distribution. Thus, provide values for the degrees of freedom, location, and scale parameters; e.g., prior = list(sigma = c(df = 10, location = 0, scale = 5)) for a prior centered on zero with scale of 5 and 10 degrees of freedom. The half-Student's t prior for sigma is constrained to be positive.

nu

The degrees of freedom for the Student's t likelihood model is assigned a gamma prior distribution. The default prior is prior = list(nu = c(alpha = 3, beta = 0.2)).

tau

The scale parameter for random effects, or varying intercepts, terms. This scale parameter, tau, is assigned a half-Student's t prior. To set this, use, e.g., prior = list(tau = c(df = 20, location = 0, scale = 20)).

centerx
To center predictors, provide either a logical value (TRUE, FALSE) or numeric-alike vector of length equal to the number of columns of ‘x’, where ‘numeric-alike’ means that ‘as.numeric(.)’ will be applied successfully if ‘is.numeric(.)’ is not true. Passed to scale.

prior_only
Draw samples from the prior distributions of parameters only.

chains
Number of MCMC chains to estimate.

iter
Number of samples per chain.

refresh
Stan will print the progress of the sampler every refresh number of samples; set refresh=0 to silence this.

pars
Optional; specify any additional parameters you'd like stored from the Stan model.

control
A named list of parameters to control the sampler's behavior. See stan for details. The defaults are the same rstan::stan except that adapt_delta is raised to 0.95 and max_treedepth = 15.

...
Other arguments passed to sampling. For multi-core processing, you can use cores = parallel::detectCores(), or run options(mc.cores = parallel::detectCores()) first.

Value

An object of class class geostan_fit (a list) containing:

summary

Summaries of the main parameters of interest; a data frame

diagnostic

Widely Applicable Information Criteria (WAIC) with a measure of effective number of parameters (eff_pars) and mean log pointwise predictive density (lpd), and mean residual spatial autocorrelation as measured by the Moran coefficient.

stanfit

an object of class stanfit returned by rstan::stan

data

a data frame containing the model data

family

the user-provided or default family argument used to fit the model

formula

The model formula provided by the user (not including ESF component)

slx

The slx formula

re

A list containing re, the random effects (varying intercepts) formula if provided, and Data a data frame with columns id, the grouping variable, and idx, the index values assigned to each group.

priors

Prior specifications.

x_center

If covariates are centered internally (i.e., centerx is not FALSE), then x_centers is the numeric vector of values on which the covariates were centered.

ME

The ME data list, if one was provided by the user for measurement error models.

spatial

NA, slot is maintained for use in geostan_fit methods.

Details

Fit a generalized linear model using the R formula interface. Default prior distributions are designed to be weakly informative relative to the data. Much of the functionality intended for spatial models, such as the ability to add spatially lagged covariates and observational error models, are also available in stan_glm. All of geostan's spatial models build on top of the same Stan code used in stan_glm.

Poisson models are often used to calculate incidence rates (mortality rates, or disease incidence rates) for administrative areas, like counties or census tracts; the user provided offset should be, in that case, the natural logarithm of the denominator, e.g., log-population at risk. If Y are counts of cases, and P are populations at risk, then the crude rates are Y/P, and the user should provide an offset value of log(P). For such a case, disease incidence across the collection of areas could be modeled as:

                          Y ~ Poisson(exp(offset + Mu))
                          Mu = alpha + A
                          A ~ Guass(0, tau)
                          tau ~ student(20, 0, 2),

where alpha is the mean log-risk (incidence rate) and A is a vector of (so-called) random effects, which enable partial pooling of information across observations. See the example section of this documenet for a demonstration.

Author

Connor Donegan, Connor.Donegan@UTDallas.edu

Examples

if (FALSE) {
library(ggplot2)
library(sf)
data(sentencing)

sentencing$log_e <- log(sentencing$expected_sents)
fit.pois <- stan_glm(sents ~ offset(log_e),
                     re = ~ name,
                     family = poisson(),
                     data = sentencing
 )

# MCMC diagnostics plot: Rhat values should all by very near 1
rstan::stan_rhat(fit.pois$stanfit)
 # see effective sample size for all parameters and generated quantities
 # (including residuals, predicted values, etc.)
rstan::stan_ess(fit.pois$stanfit)
# or for a particular parameter
rstan::stan_ess(fit.pois$stanfit, "alpha_re")

# Spatial autocorrelation/residual diagnostics
sp_diag(fit.pois, sentencing)

## Posterior predictive check                                        
library(bayesplot)
yrep <- posterior_predict(fit.pois, samples = 75)
y <- sentencing$sents
ppc_dens_overlay(y, yrep)
}